### 2020-05-26 #2256 @whedon
[PRE REVIEW]: OMG: A Scalable and Flexible Simulation and Testing Environment Toolbox for Intelligent Microgrid Control
**Submitting author:** @stheid (<a href="http://orcid.org/0000-0002-9461-7372">Stefan Heid</a>)
**Repository:** <a href="https://github.com/upb-lea/openmodelica-microgrid-gym" target ="_blank">https://github.com/upb-lea/openmodelica-microgrid-gym</a>
**Version:** v0.2.0
