### 2020-05-26 #2256 @danielskatz
👋 @stheid - where is the paper.md and .bib file for this submission?
