### 2020-05-26 7fdb02e
#7 solved some issues related to agents figures and the runner class
