### 2020-05-27 #2256 @stheid
Oh, i assume the issue is caused that the paper is currently not yet on the master branch. The preview worked fine. I am a bit worried to make an ad-hoc merge. Please give me until the end of this work-week to properly finalize our development branch and merge to master.
