### 2020-05-27 d37458c
#12 improved plotting example
