### 2020-05-27 635b0cf
Updated banner in user guide amd readme
