### 2020-05-27 66fcc50
#12 added userguide description reorganized basic examples
