+++ docs/index.rst
@@ -21,7 +21,7 @@ For the creation of additional grid constellations, Openmodelica (nightly build 
    :titlesonly:
    :caption: User Guide:
 
-   parts/user_guide/user_guide
+   parts/user_guide/getting_started
    parts/user_guide/OpenModelica
    parts/user_guide/fmu
    parts/user_guide/Pythoncode
