### 2020-05-27 89d5a19
#12 redesigned example to new plot template
