+++ examples/single_inverter_current_control_safe_opt.py
@@ -17,6 +17,7 @@ from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
 from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
 # Choose which controller parameters should be adjusted by SafeOpt.
@@ -33,7 +34,7 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
 max_episode_steps = 300  # number of simulation steps per episode
-num_episodes = 10  # number of simulation episodes (i.e. SafeOpt iterations)
+num_episodes = 5  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -186,15 +187,26 @@ if __name__ == '__main__':
     # Using an inverter supplying a load
     # - using the reward function described above as callable in the env
     # - viz_cols used to choose which measurement values should be displayed (here, only the 3 currents across the
-    #   inductors of the inverters are plotted. Alternatively, the grid frequency and the currents transformed by
-    #   the agent to the dq0 system can be plotted if requested.)
+    #   inductors of the inverters are plotted. Labels and grid is adjusted using the PlotTmpl (For more information,
+    #   see UserGuide)
     # - inputs to the models are the connection points to the inverters (see user guide for more details)
     # - model outputs are the the 3 currents through the inductors and the 3 voltages across the capacitors
+
+    def xylables(fig):
+        ax = fig.gca()
+        ax.set_xlabel(r'$t\,/\,\mathrm{ms}$')
+        ax.set_ylabel('$i_{\mathrm{abc}}\,/\,\mathrm{A}$')
+        ax.grid(which='both')
+        #fig.savefig('Inductor_currents.pdf')
+
     env = gym.make('openmodelica_microgrid_gym:ModelicaEnv_test-v1',
                    reward_fun=Reward().rew_fun,
                    time_step=delta_t,
-                   # viz_cols=['master.freq', 'master.CVI*', 'lc1.ind*'],
-                   viz_cols=['lc1.ind*'],
+                   viz_cols=[
+                       PlotTmpl([f'lc1.inductor{i}.i' for i in '123'],
+                                callback=xylables
+                                )
+                   ],
                    log_level=logging.INFO,
                    viz_mode='episode',
                    max_episode_steps=max_episode_steps,
@@ -224,15 +236,12 @@ if __name__ == '__main__':
     # Show best episode measurment (current) plot
     best_env_plt = runner.best_episode['best_env_plt']
     ax = best_env_plt[0].axes[0]
-    ax.set_xlabel(r'$t\,/\,\mathrm{ms}$')
-    ax.set_ylabel('$i_{\mathrm{abc}}\,/\,\mathrm{A}$')
-    ax.grid(which='both')
     ax.set_title('Best Episode')
     best_env_plt[0].show()
     best_env_plt[0].savefig('best_env_plt.png')
 
     # Show last performance plot
-    best_agent_plt = runner.agent.figure
+    best_agent_plt = runner.best_episode['agent_plt']
[…]
