### 2020-05-27 c9b13ac
#12 show env plot after each episode
