### 2020-05-27 90d4abb
#12 save last agentplot
