### 2020-05-27 9a12e7b
minor change in visualization of badges in readme (#34)
