### 2020-05-27 90d82f2
#12 renamed dict and keys to save best_env_episode and last agent plot
