### 2020-05-27 29a8c5b
Feature/12 new plots (#33)
* refactored packages

* some steps towards automatic merging of parameters in the plottmpl

*  implemented Plottemplate fixed test and improved coverage

* added documentation for plottmpl and flatten_together

* integrated usage into modelica.py and added plotting example

* solved some issues related to agents figures and the runner class

* improved plotting example

* best episode API docu

* added userguide description
reorganized basic examples

* redesigned example to new plot template

* show env plot after each episode

* save last agentplot

* renamed dict
[…]
