+++ openmodelica_microgrid_gym/env/__init__.py
@@ -1,2 +1,4 @@
-from .modelica import *
-from .recorder import *
+from .modelica import ModelicaEnv
+from .plot import PlotTmpl
+
+__all__ = ['ModelicaEnv', 'PlotTmpl']
