### 2020-05-27 c0eeace
updated HISTORY.rst
