+++ setup.py
@@ -48,6 +48,6 @@ setup(
         "Documentation": "https://upb-lea.github.io/openmodelica-microgrid-gym/",
         "Source Code": "https://github.com/upb-lea/openmodelica-microgrid-gym",
     },
-    version='0.1.3',
+    version='0.2.0',
     zip_safe=False,
 )
