+++ _sources/api/omg.aux_ctl.pi_controllers.rst.txt
@@ -1,0 +1,7 @@
+omg.aux_ctl.pi_controllers
+====================================
+
+.. automodule:: openmodelica_microgrid_gym.aux_ctl.pi_controllers
+   :members:
+   :undoc-members:
+   :show-inheritance:
