+++ _sources/api/omg.auxiliaries.filter.rst.txt
@@ -1,7 +1,0 @@
-omg.auxiliaries.filter
-========================================
-
-.. automodule:: openmodelica_microgrid_gym.auxiliaries.filter
-   :members:
-   :undoc-members:
-   :show-inheritance:
