+++ _sources/api/omg.aux_ctl.inverter_controllers.rst.txt
@@ -1,0 +1,7 @@
+omg.aux_ctl.inverter\_controllers
+================================================
+
+.. automodule:: openmodelica_microgrid_gym.aux_ctl.inverter_contollers
+   :members:
+   :undoc-members:
+   :show-inheritance:
