+++ _sources/parts/user_guide/examples.rst.txt
@@ -1,5 +1,7 @@
-Examples
-========
+Basic Examples
+==============
+
+To get an idea how the toolbox works and how to set-up an environment, an agent or use plotting, three very basic examples are shown below. They use the network presented above, but only the first inverter will be used.
 
 
 .. toctree::
@@ -7,4 +9,21 @@ Examples
    :titlesonly:
    :glob:
 
-   examples/*
+   examples/creating_env
+   examples/basic_agent
+   examples/plotting
+
+
+
+
+Advanced Examples
+=================
+
+
+.. toctree::
+   :maxdepth: 4
+   :titlesonly:
+   :glob:
+
+   examples/single_inverter_current_control_safe_opt
+   examples/two_inverter_static_droop_control
