+++ _sources/api/omg.aux_ctl.base.rst.txt
@@ -1,0 +1,7 @@
+omg.aux_ctl.base
+======================================
+
+.. automodule:: openmodelica_microgrid_gym.aux_ctl.base
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
