### 2020-05-27 9374076
Update HISTORY.rst
