+++ HISTORY.rst
@@ -2,8 +2,8 @@
 History
 =======
 
-2.0.0
----------------
+0.2.0 (2020-05-27)
+------------------
 
 
 API
