+++ HISTORY.rst
@@ -7,7 +7,7 @@ History
 
 
 API
-```
+^^^
 * ModelicaEnv:
    - reward function parameter
    - vis_cols now also supports Plotting templates
@@ -19,15 +19,15 @@ API
 * Runner: plotting can be disabled
 
 Examples
-````````
+^^^^^^^^
 * added example for plotting
 
 Performance
-```````````
+^^^^^^^^^^^
 * 6.6× speedup
 
 Dependencies
-````````````
+^^^^^^^^^^^^
 * Increased Language Level to Python 3.8
 
 
