### 2020-05-28 #2256 @stheid
@danielskatz i assume it is fine now. the whedon.theoj preview produces a paper if i omit specifying the branch (meaning master branch)
