### 2020-05-28 #2256 @stheid
![image](https://user-images.githubusercontent.com/2736207/83120700-aaa88000-a0d1-11ea-8566-077a97f2152d.png)

And is there a way to fix the missing whitespace typo? I assume everything will be automatic from here on so the typo would also be on you website if this review comes to an sucessfull end.
