### 2020-05-29 d0a5c94
fixed episode length in README.rst
