+++ README.rst
@@ -94,6 +94,7 @@ The environment is initialized and run like any other OpenAI Gym
 
     if __name__ == '__main__':
         env = gym.make('openmodelica_microgrid_gym:ModelicaEnv-v1',
+                       max_episode_steps=None,
                        model_input=['i1p1', 'i1p2', 'i1p3'],
                        model_output=dict(lc1=['inductor1.i', 'inductor2.i', 'inductor3.i']),
                        model_path='grid.network.fmu')
