### 2020-06-08 b902b62
first try with feature engeneering
