+++ examples/single_inverter_voltage_current_control_safe_opt.py
@@ -5,6 +5,7 @@
 
 
 import logging
+from time import strftime, gmtime
 from typing import List
 
 import GPy
@@ -16,7 +17,8 @@ import matplotlib.pyplot as plt
 from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
-from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController, \
+    MultiPhaseDQ0PIPIController
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
@@ -34,7 +36,7 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
 max_episode_steps = 300  # number of simulation steps per episode
-num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
+num_episodes = 30  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -54,7 +56,8 @@ class Reward:
         if self._idx is None:
             self._idx = nested_map(
                 lambda n: obs.index(n),
-                [[f'lc1.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0']])
+                [[f'lc1.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0'],
+                 [f'lc1.capacitor{k}.v' for k in '123'], [f'master.SPV{k}' for k in 'dq0']])
 
     def rew_fun(self, cols: List[str], data: np.ndarray) -> float:
         """
@@ -72,10 +75,14 @@ class Reward:
 
         Iabc_master = data[idx[0]]  # 3 phase currents at LC inductors
         phase = data[idx[1]]  # phase from the master controller needed for transformation
+        Vabc_master = data[idx[3]]  # 3 phase currents at LC inductors
 
         # setpoints
         ISPdq0_master = data[idx[2]]  # setting dq reference
         ISPabc_master = dq0_to_abc(ISPdq0_master, phase)  # convert dq set-points into three-phase abc coordinates
+        VSPdq0_master = data[idx[4]]  # setting dq reference
+        VSPabc_master = dq0_to_abc(VSPdq0_master, phase)  # convert dq set-points into three-phase abc coordinates
+
 
         # control error = mean-root-error (MRE) of reference minus measurement
         # (due to normalization the control error is often around zero -> compared to MSE metric, the MRE provides
@@ -83,8 +90,11 @@ class Reward:
         # plus barrier penalty for violating the current constraint
         error = np.sum((np.abs((ISPabc_master - Iabc_master)) / iLimit) ** 0.5, axis=0) \
                 + -np.sum(mu * np.log(1 - np.maximum(np.abs(Iabc_master) - iNominal, 0) / (iLimit - iNominal)), axis=0) \
-                * max_episode_steps
+                * max_episode_steps \
+                + np.sum((np.abs((VSPabc_master - Vabc_master)) / nomVoltPeak) ** 0.5, axis=0)
 
+        if np.isnan(error):
+            asd=1
         return -error.squeeze()
 
 
[…]
