+++ examples/two_inverter_droop_safe_opt.py
@@ -5,6 +5,8 @@
 
 
 import logging
+from functools import partial
+from time import strftime, gmtime
 from typing import List
 
 import GPy
@@ -16,7 +18,8 @@ import matplotlib.pyplot as plt
 from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
-from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController, \
+    MultiPhaseDQ0PIPIController, PLLParams, InverseDroopParams, MultiPhaseDQCurrentController
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
@@ -25,16 +28,16 @@ from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 # - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
 # - Kpi: 2D example: Kp and Ki are adjusted simultaneously
 
-adjust = 'Kpi'
+adjust = 'pDroop'
 
 # Check if really only one simulation scenario was selected
-if adjust not in {'Kp', 'Ki', 'Kpi'}:
-    raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
+if adjust not in {'pDroop', 'Ki', 'all'}:
+    raise ValueError("Please set 'adjust' to one of the following values: 'pDroop', 'Ki', 'all'")
 
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
-max_episode_steps = 300  # number of simulation steps per episode
-num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
+max_episode_steps = 6000  # number of simulation steps per episode
+num_episodes = 20  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -45,6 +48,16 @@ DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
 i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
 
+def load_step(t, gain):
+    """
+    Defines a load step after 0.2 s
+    Doubles the load parameters
+    :param t:
+    :param gain: device parameter
+    :return: Dictionary with load parameters
+    """
+    return 1*gain if t < .15 else 5*gain
+
 
 class Reward:
     def __init__(self):
@@ -54,7 +67,9 @@ class Reward:
         if self._idx is None:
             self._idx = nested_map(
                 lambda n: obs.index(n),
-                [[f'lc1.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0']])
+                [f'slave.freq'])
+                #[[f'lc1.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0'],
+                # [f'lc1.capacitor{k}.v' for k in '123'], [f'master.SPV{k}' for k in 'dq0']])
 
[…]
