+++ .travis.yml
@@ -1,7 +1,7 @@
 language: python
 
 python:
-  - "3.8" # Travis Python is actually not used since we use conda, but this keeps it organized.
+  - "3.6" # Travis Python is actually not used since we use conda, but this keeps it organized.
 
 install:
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
