### 2020-06-11 #2256 @labarba
👋 @sjpfenninger — Could you have a look at this submission and let us know if you can handle it as editor?
Title: OMG: A Scalable and Flexible Simulation and Testing Environment Toolbox for Intelligent Microgrid Control
[ Article proof](https://github.com/openjournals/joss-papers/blob/joss.02256/joss.02256/10.21105.joss.02256.pdf)
