### 2020-06-15 fc1e239
- added noise measurement - created physical testbench env (connect to FPGA via SSH) - generated hardware runner - added noise measurement - added SPs to agent plots - added loadstep to prepare MC - set initial parameters e.g. R,L,... using self.self.model_parameters ind env.reset() - shifted initial parameter setting from modelica.py to pyfmi.py reset - added flexible save folder - include episodic agent to safeopt agent + Change performance to property (getter and setter) - generate MC runner, which runs an additional loop for n_MC Monte-Carlo experiments per episode averaging the performanc
[…]
