+++ openmodelica_microgrid_gym/env/plot.py
@@ -82,7 +82,8 @@ class PlotTmpl:
         """
         if self._callback is not None:
             self._callback(fig)
-        plt.show()
+        else:
+            plt.show()
 
     def __iter__(self):
         self.i = -1
