+++ openmodelica_microgrid_gym/agents/safeopt.py
@@ -9,17 +9,19 @@ from GPy.kern import Kern
 from matplotlib.figure import Figure
 from safeopt import SafeOptSwarm
 
+from openmodelica_microgrid_gym.agents.episodic import EpisodicLearnerAgent
 from openmodelica_microgrid_gym.agents.staticctrl import StaticControlAgent
 from openmodelica_microgrid_gym.agents.util import MutableParams
-from openmodelica_microgrid_gym.aux_ctl import Controller
+from openmodelica_microgrid_gym.aux_ctl import Controller, DDS
 
 logger = logging.getLogger(__name__)
 
 
-class SafeOptAgent(StaticControlAgent):
+class SafeOptAgent(StaticControlAgent, EpisodicLearnerAgent):
     def __init__(self, mutable_params: Union[dict, list], abort_reward: int, kernel: Kern, gp_params: Dict[str, Any],
                  ctrls: List[Controller],
                  obs_template: Mapping[str, List[Union[List[str], np.ndarray]]], obs_varnames: List[str] = None,
+                 min_performance: float = None,
                  **kwargs):
         """
         Agent to execute safeopt algorithm (https://arxiv.org/abs/1509.01066) to control the environment by using
@@ -46,6 +48,7 @@ class SafeOptAgent(StaticControlAgent):
             The values will be passed as parameters to the controllers step function.
         :param obs_varnames: list of variable names that match the values of the observations
          passed in the act function. Will be automatically set by the Runner class
+         :param min_performance: Minial allowed performance to define parameters as safe
         """
         self.params = MutableParams(
             list(mutable_params.values()) if isinstance(mutable_params, dict) else mutable_params)
@@ -57,11 +60,15 @@ class SafeOptAgent(StaticControlAgent):
         self.explore_threshold = gp_params['explore_threshold']
 
         self.abort_reward = abort_reward
-        self.episode_reward = None
+        self.episode_return = None
         self.optimizer = None
-        self.inital_performance = None
-        self.last_best_performance = None
-        self.performance = None
+        self._performance = None
+        self._min_performance = min_performance
+        self.initial_performance = min_performance/2
+        self.last_best_performance = 0      # set to 0 due to in MC otherwise we do not get the best/worst before the
+        # first update_params after the MC loop
+        self.last_worst_performance = 0
+        self.unsafe = False
 
         self._iterations = 0
 
@@ -83,10 +90,12 @@ class SafeOptAgent(StaticControlAgent):
 
         self.params.reset()
         self.optimizer = None
-        self.episode_reward = 0
-        self.inital_performance = None
-        self.last_best_performance = None
-        self.performance = None
+        self.episode_return = 0
+        self.initial_performance = 1
+        self._performance = None
+        self.last_best_performance = 0
+        self.last_worst_performance = 0
+        self.unsafe = False
 
         self._iterations = 0
 
@@ -102,10 +111,14 @@ class SafeOptAgent(StaticControlAgent):
         :return:
         """
         self._iterations += 1
-        self.episode_reward += reward or 0
+        self.episode_return += reward or 0
         if terminated:
-            # calculate MSE, divide Summed error by length of measurement
-            self.performance = self.episode_reward / self._iterations
+
+            if self.optimizer is None:
+                self.initial_performance = self.episode_return
+
+            #self.performance = self._iterations / (self.episode_return * self.initial_performance)
+            self.performance = (self.episode_return - self._min_performance) / (self.initial_performance - self.min_performance)
             # safeopt update step
             self.update_params()
             # reset for new episode
@@ -118,12 +131,8 @@ class SafeOptAgent(StaticControlAgent):
         """
         if self.optimizer is None:
             # First Iteration
-            self.inital_performance = self.performance
-
[…]
