+++ openmodelica_microgrid_gym/agents/safeopt_remote.py
@@ -1,0 +1,197 @@
+import importlib
+import logging
+from typing import Dict, Union, Any, List, Mapping
+
+import GPy
+import matplotlib.pyplot as plt
+import numpy as np
+from GPy.kern import Kern
+from matplotlib.figure import Figure
+from safeopt import SafeOptSwarm
+
+from openmodelica_microgrid_gym import Agent
+from openmodelica_microgrid_gym.agents.util import MutableParams
+from openmodelica_microgrid_gym.aux_ctl import Controller
+
+
+logger = logging.getLogger(__name__)
+
+
+class remote_SafeOptAgent(Agent):
+    def __init__(self, mutable_params: Union[dict, list], abort_reward: int, kernel: Kern, gp_params: Dict[str, Any],
+                 ctrls: List[Controller],
+                 obs_template: Mapping[str, List[Union[List[str], np.ndarray]]], obs_varnames: List[str] = None,
+                 **kwargs):
+        """
+        Agent to execute safeopt algorithm (https://arxiv.org/abs/1509.01066) to control the environment by using
+        auxiliary controllers and Gaussian process to adopt the controller parameters (mutable_params) to safely
+        increase the performance.
+
+        :param mutable_params: safe inital controller parameters to adopt
+        :param abort_reward: factor to multiply with the initial reward to give back an abort_reward-times higher
+               negative reward in case of limit exceeded
+        :param kernel: kernel for the Gaussian process unsing GPy
+        :param gp_params: kernel parameters like bounds and lengthscale
+        :param ctrls: Controllers that are feed with the observations and exert actions on the environment
+        :param obs_template:
+            Template describing how the observation array should be transformed and passed to the internal controllers.
+            The key must match the name field of an internal controller.
+            The values are a list of:
+                - list of strings
+                    - matching variable names of the state
+                    - must match self.obs_varnames
+                    - will be substituted by the values on runtime
+                    - will be passed as an np.array of floats to the controller
+                - np.ndarray of floats (to be passed statically to the controller)
+                - a mixture of static and dynamic values in one parameter is not supported for performance reasons.
+            The values will be passed as parameters to the controllers step function.
+        :param obs_varnames: list of variable names that match the values of the observations
+         passed in the act function. Will be automatically set by the Runner class
+        """
+        self.params = MutableParams(
+            list(mutable_params.values()) if isinstance(mutable_params, dict) else mutable_params)
+        self.kernel = kernel
+        self.bounds = gp_params['bounds']
+        self.noise_var = gp_params['noise_var']
+        self.prior_mean = gp_params['prior_mean']
+        self.safe_threshold = gp_params['safe_threshold']
+        self.explore_threshold = gp_params['explore_threshold']
+
+        self.abort_reward = abort_reward
+        self.episode_reward = None
+        self.optimizer = None
+        self.inital_performance = None
+        self.last_best_performance = None
+        self.performance = None
+
+        self._iterations = 0
+
+        super().__init__(ctrls, obs_template, obs_varnames, **kwargs)
+        self.history.cols = ['J', 'Params']
+
+    def reset(self):
+        """
+        Resets the kernel, episodic reward and the optimizer
+        """
+        # reinstantiate kernel
+
+        kernel_params = self.kernel.to_dict()
+        cls_name = kernel_params['class']
+        mod = importlib.import_module('.'.join(cls_name.split('.')[:-1]))
+        cls = getattr(mod, cls_name.split('.')[-1])
+        remaining_params = {k: v for k, v in kernel_params.items() if k not in {'class', 'useGPU'}}
+        self.kernel = cls(**remaining_params)
+
+        self.params.reset()
+        self.optimizer = None
+        self.episode_reward = 0
+        self.inital_performance = None
[…]
