+++ examples/testbench_safeopt.py
@@ -12,9 +12,11 @@ import gym
 import numpy as np
 
 import matplotlib.pyplot as plt
+from tqdm import tqdm
 
 from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
+from openmodelica_microgrid_gym.agents.safeopt_remote import remote_SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
 from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
 from openmodelica_microgrid_gym.env import PlotTmpl
@@ -25,6 +27,7 @@ from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 # - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
 # - Kpi: 2D example: Kp and Ki are adjusted simultaneously
 
+
 adjust = 'Kpi'
 
 # Check if really only one simulation scenario was selected
@@ -34,8 +37,8 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
 max_episode_steps = 300  # number of simulation steps per episode
-num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
-v_DC = 65  # DC-link voltage / V; will be set as model parameter in the FMU
+num_episodes = 5  # number of simulation episodes (i.e. SafeOpt iterations)
+v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
 iLimit = 30  # inverter current limit / A
@@ -87,7 +90,6 @@ class Reward:
 
         return -error.squeeze()
 
-
 if __name__ == '__main__':
     #####################################
     # Definitions for the GP
@@ -134,42 +136,26 @@ if __name__ == '__main__':
     mutable_params = None
     current_dqp_iparams = None
     if adjust == 'Kp':
-        # mutable_params = parameter (Kp gain of the current controller of the inverter) to be optimized using
-        # the SafeOpt algorithm
-        mutable_params = dict(currentP=MutableFloat(5e-3))
-
-        # Define the PI parameters for the current controller of the inverter
-        current_dqp_iparams = PI_params(kP=mutable_params['currentP'], kI=115, limits=(-1, 1))
+        # set inital values for
+        Kp = 5e-3
 
     # For 1D example, if Ki should be adjusted
     elif adjust == 'Ki':
-        mutable_params = dict(currentI=MutableFloat(10))
-        current_dqp_iparams = PI_params(kP=10e-3, kI=mutable_params['currentI'], limits=(-1, 1))
+        Ki = 10
 
     # For 2D example, choose Kp and Ki as mutable parameters
     elif adjust == 'Kpi':
-        mutable_params = dict(currentP=MutableFloat(10e-3), currentI=MutableFloat(10))
-        current_dqp_iparams = PI_params(kP=mutable_params['currentP'], kI=mutable_params['currentI'], limits=(-1, 1))
+        Kp = 5e-3
+        Ki = 10
 
-    # Define the droop parameters for the inverter of the active power Watt/Hz (DroopGain), delta_t (0.005) used for the
-    # filter and the nominal frequency
-    # Droop controller used to calculate the virtual frequency drop due to load changes
-    droop_param = DroopParams(DroopGain, 0.005, nomFreq)
 
-    # Define the Q-droop parameters for the inverter of the reactive power VAR/Volt, delta_t (0.002) used for the
-    # filter and the nominal voltage
[…]
