### 2020-06-19 #2256 @stheid
> Hey @stheid - did you have any recommendations for reviewers (list [here](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822))? If not, I will assign a couple. Thanks!

ahm, i have no specific preference. i guess someone with the background to electrical engeneering; control; or reinforcement learning would be good
