### 2020-06-30 a4845ca
First skeleton of the new network configuration file and coresponding classes
