### 2020-06-30 d48dff7
snapshot before refactoring
