+++ openmodelica_microgrid_gym/net/net.py
@@ -1,30 +1,35 @@
 import sys
+from itertools import chain
 from typing import List, Optional
 
+import numpy as np
 import yaml
 
-from openmodelica_microgrid_gym import ModelicaEnv
 from openmodelica_microgrid_gym.aux_ctl import PLL, PLLParams
 
 
 class Network:
-    def __init__(self, ts, v_nom, freq_nom, i_nom, i_lim, v_DC=1000):
+    def __init__(self, ts, v_nom, freq_nom=50):
         self.ts = ts
         self.v_nom = v_nom
         self.freq_nom = freq_nom
-        self.i_nom = i_nom
-        self.i_lim = i_lim
-        self.v_DC = v_DC
         self.components = []  # type: List[Component]
 
     @classmethod
-    def load(cls, configurl):
+    def load(cls, configurl='net.yaml'):
+        """
+        Initialize object from config file
+
+        :param configurl:
+        :return:
+        """
         data = yaml.safe_load(open(configurl))
         components = data['components']
         del data['components']
         self = cls(**data)
 
         for name, component in components.items():
+            # resolve class from 'cls' argument
             comp_cls = component['cls']
             del component['cls']
             comp_cls = getattr(sys.modules[__name__], comp_cls)
@@ -35,28 +40,62 @@ class Network:
             if 'out' in component:
                 component['out_vars'] = component.pop('out')
 
-            self.components.append(comp_cls(net=self, **component))
+            # instanciate component class
+            try:
+                self.components.append(comp_cls(net=self, **component))
+            except AttributeError as e:
+                raise AttributeError(f'{e!s}, please validate {configurl}')
 
         return self
 
-    def statekeys(self, keys):
+    def augment_actions(self, actions):
+        """
+        Allows the network to add additional parameters like changing loads to the simulation
+
+        :param actions:
+        :return: mapping of additional actions and list of actions.
+        """
+        # TODO get params from components
+        return {}, actions
+
+    def augment_state(self, state: np.ndarray) -> np.ndarray:
+        """
+        Allows the network to provide additional output variables in order to provide measurements and reference
+        information the RL agent needs to understand its rewards
+        :param state:
+        :return:
+        """
+        # TODO calculate additional outputs
+        return np.hstack([comp.normalized_outputs(state) for comp in self.components])
+
+
+
+
+
+
+    def _statekeys(self, keys):
         for comp in self.components:
             comp.set_outidx(keys)
 
-    def getstatekeys(self):
-        return [comp.outputnames() for comp in self.components]
 
-    def augment(self, state):
-        return np.hstack([comp.normalized_outputs(state) for comp in self.components])
+    def in_keys(self):
+        return [comp.inkeys(onlysrc) for comp in self.components]
+
+    def out_keys(self, onlysrc=False):
+        return [comp.outkeys(onlysrc) for comp in self.components]
 
 
 class Component:
     def __init__(self, net: Network, id, in_vars=None, out_vars=None):
         self.net = net
         self.id = id
+        for attr in chain.from_iterable((f.keys() for f in filter(None, (in_vars, out_vars)))):
+            if not hasattr(self, attr):
+                raise AttributeError(f'{self.__class__} no such attribute: {attr}')
         self.in_vars = in_vars
         self.in_idx = None  # type: Optional[dict]
 
+        self.new_outputs
[…]
