+++ openmodelica_microgrid_gym/env/modelica.py
@@ -402,19 +402,21 @@ class ModelicaEnv(gym.Env):
 
 
 class NormalizedEnv(ModelicaEnv):
-    def __init__(self, netfile, **kwds):
-        self.net = Network.load(netfile)
-        super().__init__(time_step=self.net.ts, **kwds)
+    def __init__(self, net, **kwds):
+        self.net = Network.load(net)
+        super().__init__(time_step=self.net.ts, model_input=self.net.in_keys(True), model_output=self.net.out_keys(True),
+                         **kwds)
         self.net.statekeys(self.model_output_names)
 
     def reset(self) -> np.ndarray:
-        obs =super().reset()
-        outputs = self.net.normalizedoutput(obs)
+        obs = super().reset()
+        outputs = self.net.augment(obs)
         return outputs
 
     def step(self, action: Sequence) -> Tuple[np.ndarray, float, bool, Mapping]:
         params, actions = self.net.input(action)
-        self.model.set(**params)
-        obs,rew,done, info = super().step(actions)
-        outputs = self.net.normalizedoutput(obs)
-        return outputs,rew,done,info
+        if params:
+            self.model.set(**params)
+        obs, rew, done, info = super().step(actions)
+        outputs = self.net.augment(obs)
+        return outputs, rew, done, info
