### 2020-07-01 cdb6802
# 17 creat fully commented examples for 1 inv voltage and current control
