+++ examples/single_inverter_voltage_current_control_safe_opt.py
@@ -1,7 +1,7 @@
 #####################################
 # Example using a FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
-# Simulation setup: Single inverter supplying 15 A d-current to an RL-load via a LC filter
-# Controller: PI current controller gain parameters are optimized by SafeOpt
+# Simulation setup: Single voltage forming inverter supplying an RL-load via an LC filter
+# Controller: Cascaded PI-PI voltage and current controller gain parameters are optimized by SafeOpt
 
 
 import logging
@@ -22,16 +22,6 @@ from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhas
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
-# Choose which controller parameters should be adjusted by SafeOpt.
-# - Kp: 1D example: Only the proportional gain Kp of the PI controller is adjusted
-# - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
-# - Kpi: 2D example: Kp and Ki are adjusted simultaneously
-
-adjust = 'Kpi'
-
-# Check if really only one simulation scenario was selected
-if adjust not in {'Kp', 'Ki', 'Kpi'}:
-    raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
@@ -45,7 +35,6 @@ iNominal = 20  # nominal inverter current / A
 mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
-i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
 
 
 class Reward:
@@ -61,10 +50,11 @@ class Reward:
 
     def rew_fun(self, cols: List[str], data: np.ndarray) -> float:
         """
-        Defines the reward function for the environment. Uses the observations and setpoints to evaluate the quality of the
-        used parameters.
-        Takes current measurement and setpoints so calculate the mean-root-error control error and uses a logarithmic
-        barrier function in case of violating the current limit. Barrier function is adjustable using parameter mu.
+        Defines the reward function for the environment. Uses the observations and setpoints to evaluate the quality of
+        the used parameters.
+        Takes current and voltage measurement and setpoints to calculate the mean-root-error control error and uses a
+        logarithmic barrier function in case of violating the current limit. Barrier function is adjustable using
+        parameter mu.
 
         :param cols: list of variable names of the data
         :param data: observation data from the environment (ControlVariables, e.g. currents and voltages)
@@ -78,9 +68,9 @@ class Reward:
         Vabc_master = data[idx[3]]  # 3 phase currents at LC inductors
 
         # setpoints
-        ISPdq0_master = data[idx[2]]  # setting dq reference
+        ISPdq0_master = data[idx[2]]  # setting dq current reference
         ISPabc_master = dq0_to_abc(ISPdq0_master, phase)  # convert dq set-points into three-phase abc coordinates
-        VSPdq0_master = data[idx[4]]  # setting dq reference
[…]
