+++ examples/two_inverter_droop_safe_opt.py
@@ -1,7 +1,7 @@
 #####################################
 # Example using a FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
-# Simulation setup: Single inverter supplying 15 A d-current to an RL-load via a LC filter
-# Controller: PI current controller gain parameters are optimized by SafeOpt
+# Simulation setup: Two inverters supplying a RL-load via a LC filter
+# Controller: droop controller gains are optimized by SafeOpt
 
 
 import logging
@@ -23,21 +23,10 @@ from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhas
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
-# Choose which controller parameters should be adjusted by SafeOpt.
-# - Kp: 1D example: Only the proportional gain Kp of the PI controller is adjusted
-# - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
-# - Kpi: 2D example: Kp and Ki are adjusted simultaneously
-
-adjust = 'pDroop'
-
-# Check if really only one simulation scenario was selected
-if adjust not in {'pDroop', 'Ki', 'all'}:
-    raise ValueError("Please set 'adjust' to one of the following values: 'pDroop', 'Ki', 'all'")
-
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
 max_episode_steps = 6000  # number of simulation steps per episode
-num_episodes = 20  # number of simulation episodes (i.e. SafeOpt iterations)
+num_episodes = 30  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -46,7 +35,7 @@ iNominal = 20  # nominal inverter current / A
 mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
-i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+
 
 def load_step(t, gain):
     """
@@ -56,7 +45,7 @@ def load_step(t, gain):
     :param gain: device parameter
     :return: Dictionary with load parameters
     """
-    return 1*gain if t < .15 else 5*gain
+    return 1 * gain if t < .15 else 5 * gain
 
 
 class Reward:
@@ -67,9 +56,8 @@ class Reward:
         if self._idx is None:
             self._idx = nested_map(
                 lambda n: obs.index(n),
-                [f'slave.freq'])
-                #[[f'lc1.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0'],
-                # [f'lc1.capacitor{k}.v' for k in '123'], [f'master.SPV{k}' for k in 'dq0']])
+                [[f'slave.freq'],
+                 [f'master.CVV{s}' for s in 'dq0']])
 
     def rew_fun(self, cols: List[str], data: np.ndarray) -> float:
         """
@@ -86,31 +74,14 @@ class Reward:
         idx = self._idx
         freq = data[idx[0]]
 
-        """
-        Iabc_master = data[idx[0]]  # 3 phase currents at LC inductors
-        phase = data[idx[1]]  # phase from the master controller needed for transformation
-        Vabc_master = data[idx[3]]  # 3 phase currents at LC inductors
-
-        # setpoints
-        ISPdq0_master = data[idx[2]]  # setting dq reference
[…]
