### 2020-07-01 caa3dc0
normalization and proper reference value calculation is missing. The helper functions arround should be complete now.
