+++ openmodelica_microgrid_gym/net/net.py
@@ -4,6 +4,8 @@ from typing import List, Optional
 
 import numpy as np
 import yaml
+from more_itertools import flatten, collapse
+import numexpr as ne
 
 from openmodelica_microgrid_gym.aux_ctl import PLL, PLLParams
 
@@ -13,7 +15,6 @@ class Network:
         self.ts = ts
         self.v_nom = v_nom
         self.freq_nom = freq_nom
-        self.components = []  # type: List[Component]
 
     @classmethod
     def load(cls, configurl='net.yaml'):
@@ -28,6 +29,7 @@ class Network:
         del data['components']
         self = cls(**data)
 
+        components_obj = []
         for name, component in components.items():
             # resolve class from 'cls' argument
             comp_cls = component['cls']
@@ -42,51 +44,66 @@ class Network:
 
             # instanciate component class
             try:
-                self.components.append(comp_cls(net=self, **component))
+                components_obj.append(comp_cls(net=self, **component))
             except AttributeError as e:
                 raise AttributeError(f'{e!s}, please validate {configurl}')
+        self.components = components_obj
 
         return self
 
-    def augment_actions(self, actions):
+    def params(self, actions):
         """
         Allows the network to add additional parameters like changing loads to the simulation
 
         :param actions:
         :return: mapping of additional actions and list of actions.
         """
-        # TODO get params from components
-        return {}, actions
+        d = {}
+        for comp in self.components:
+            params = comp.params(actions)
+            d.update(params)
+        return d
 
-    def augment_state(self, state: np.ndarray) -> np.ndarray:
+    def augment(self, state: np.ndarray) -> np.ndarray:
         """
         Allows the network to provide additional output variables in order to provide measurements and reference
         information the RL agent needs to understand its rewards
         :param state:
         :return:
         """
-        # TODO calculate additional outputs
-        return np.hstack([comp.normalized_outputs(state) for comp in self.components])
+        return np.hstack([comp.augment(state) for comp in self.components])
 
+    def in_vars(self):
+        return list(collapse([comp.get_in_vars() for comp in self.components]))
 
+    def out_vars(self, with_aug=True, flattened=True):
+        r = [comp.get_out_vars(with_aug) for comp in self.components]
+        if flattened:
+            return list(collapse(r))
+        return r
 
+    @property
+    def components(self):
+        return self._components
 
-
-
-    def _statekeys(self, keys):
+    @components.setter
+    def components(self, val):
+        self._components = val
+        keys = self.out_vars(with_aug=False, flattened=True)
         for comp in self.components:
             comp.set_outidx(keys)
 
 
-    def in_keys(self):
-        return [comp.inkeys(onlysrc) for comp in self.components]
-
-    def out_keys(self, onlysrc=False):
-        return [comp.outkeys(onlysrc) for comp in self.components]
-
-
 class Component:
-    def __init__(self, net: Network, id, in_vars=None, out_vars=None):
+    def __init__(self, net: Network, id=None, in_vars=None, out_vars=None, out_calc=None):
+        """
+
+        :param net:
+        :param id:
+        :param in_vars:
+        :param out_vars:
+        :param out_calc: mapping from attr name to
+        """
         self.net = net
         self.id = id
         for attr in chain.from_iterable((f.keys() for f in filter(None, (in_vars, out_vars)))):
@@ -95,31 +112,34 @@ class Component:
         self.in_vars = in_vars
         self.in_idx = None  # type: Optional[dict]
 
-        self.new_outputs
[…]
