### 2020-07-02 c45d516
fixed x axis dimension included timestep to d.findex added start_time
