+++ examples/single_inverter_voltage_current_control_safe_opt.py
@@ -1,6 +1,6 @@
 #####################################
-# Example using a FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
-# Simulation setup: Single voltage forming inverter supplying an RL-load via an LC filter
+# Example using an FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
+# Simulation setup: Single voltage forming inverter supplying an RL-load via an LC-filter
 # Controller: Cascaded PI-PI voltage and current controller gain parameters are optimized by SafeOpt
 
 
@@ -63,24 +63,23 @@ class Reward:
         self.set_idx(cols)
         idx = self._idx
 
-        Iabc_master = data[idx[0]]  # 3 phase currents at LC inductors
+        iabc_master = data[idx[0]]  # 3 phase currents at LC inductors
         phase = data[idx[1]]  # phase from the master controller needed for transformation
-        Vabc_master = data[idx[3]]  # 3 phase currents at LC inductors
+        vabc_master = data[idx[3]]  # 3 phase currents at LC inductors
 
-        # setpoints
-        ISPdq0_master = data[idx[2]]  # setting dq current reference
-        ISPabc_master = dq0_to_abc(ISPdq0_master, phase)  # convert dq set-points into three-phase abc coordinates
-        VSPdq0_master = data[idx[4]]  # setting dq voltage reference
-        VSPabc_master = dq0_to_abc(VSPdq0_master, phase)  # convert dq set-points into three-phase abc coordinates
-
+        # set points (sp)
+        isp_dq0_master = data[idx[2]]  # setting dq current reference
+        isp_abc_master = dq0_to_abc(isp_dq0_master, phase)  # convert dq set-points into three-phase abc coordinates
+        vsp_dq0_master = data[idx[4]]  # setting dq voltage reference
+        vsp_abc_master = dq0_to_abc(vsp_dq0_master, phase)  # convert dq set-points into three-phase abc coordinates
 
         # control error = mean-root-error (MRE) of reference minus measurement
         # (due to normalization the control error is often around zero -> compared to MSE metric, the MRE provides
         #  better, i.e. more significant,  gradients)
         # plus barrier penalty for violating the current constraint
-        error = np.sum((np.abs((ISPabc_master - Iabc_master)) / iLimit) ** 0.5, axis=0) \
-                + -np.sum(mu * np.log(1 - np.maximum(np.abs(Iabc_master) - iNominal, 0) / (iLimit - iNominal)), axis=0) \
-                + np.sum((np.abs((VSPabc_master - Vabc_master)) / nomVoltPeak) ** 0.5, axis=0)
+        error = np.sum((np.abs((isp_abc_master - iabc_master)) / iLimit) ** 0.5, axis=0) \
+                + -np.sum(mu * np.log(1 - np.maximum(np.abs(iabc_master) - iNominal, 0) / (iLimit - iNominal)), axis=0)\
+                + np.sum((np.abs((vsp_abc_master - vabc_master)) / nomVoltPeak) ** 0.5, axis=0)
 
         return -error.squeeze()
 
@@ -92,14 +91,13 @@ if __name__ == '__main__':
     noise_var = 0.001 ** 2  # measurement noise sigma_omega
     prior_var = 2  # prior variance of the GP
 
-
     # Choose Kp and Ki (current and voltage controller) as mutable parameters (below) and define bounds and lengthscale
     # for both of them
     bounds = [(0.0, 0.03), (0, 300),(0.0, 0.03), (0, 300)] # bounds on the input variable current-Ki&Kp and voltage-Ki&Kp
[…]
