### 2020-07-02 2205f95
normalization and inverted droop for slave controller
