### 2020-07-04 #2435 @whedon
[REVIEW]: OMG: A Scalable and Flexible Simulation and TestingEnvironment Toolbox for Intelligent Microgrid Control
**Submitting author:** @stheid (<a href="http://orcid.org/0000-0002-9461-7372">Stefan Heid</a>)
**Repository:** <a href="https://github.com/upb-lea/openmodelica-microgrid-gym" target ="_blank">https://github.com/upb-lea/openmodelica-microgrid-gym</a>
**Version:** 0.2.0-2

- [x] **Functionality:** Have the functional claims of the software been confirmed? (unable to evaluate)
- [x] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) (unable to evaluate)
