### 2020-07-08 3d860a8
# 45 added normalization bool on/off to norm env
