### 2020-07-08 8f9edea
# 45 added normalization bool on/off
