### 2020-07-08 502f33e
# 45 added normalization test env from two_inv_static_droop_control
