### 2020-07-14 8db4109
# 45 compared results to stored results from two_inverter_static_droop_controller.py -> hugh difference
