### 2020-07-14 bc38bcc
# 45 changed viz_mode from step to episode by default
