### 2020-07-17 6abb613
still contructionv parts in the grid.mo, needs to be cleaned up before pushing it somewhere else
