### 2020-07-20 be6db38
#45 fixed the main issue with the simulation not progression (was actually a bug in the config and not in the code. see #54)
