### 2020-07-20 95cf054
#45 properly handle v_DC parameters
