### 2020-07-23 #2435 @katyhuff
## katyhuff Review
This is an interesting tool for testing control system approaches for microgrids. I'm genuinely interested in testing it out further and doing more exploring with it. I think the purpose of the tool is fairly clear and the audience certainly exists. It builds cleanly from source and the pip install approach works cleanly as well. While the build succeeds, the tests fail on my MacBook (`pyfmi.fmi.FMUException: Error loading the binary. The FMU contains no binary for this platform.`). Upon reflection, the README indicated this might be the case. So, I redid this on my ubuntu machine and things went better. The tests themselves cover the majority of functions in the code, though most are more in the style of integration tests rather than unit tests. In those cases, the expected results take the form of hdf5 files and are named somewhat vaguely (test_main2, test_main3), so the purpose of these particular tests is a bit mysterious to a new user attempting to learn from them. The API has documentation is available online and there are appropriate guides for new developers and users. I tried running some of the examples, which went ok. There are a few boxes that I'm not ready to check. I'll make specific comments regarding each of those items.

### Contribution and authorship:
The submitting author (@stheid) has made major contributions to the software and the full list of paper authors seems complete. However, one entry seems to be potentially worth clarifying in the text of the paper. The contribution from Eyke Hüllermeier is not clear based on the repository history. Presumably, this individual (Eyke Hüllermeier) has contributed conceptually and intellectually. They do not seem to have committed code to the repository. This is fine, as far as I'm concerned, but should be clarified, either here in this review discussion, or potentially in the acknowledgments section of the text.

### Functionality:
If the functional claims of the software were a bit more clear in the paper, then I could better confirm them. A few of the functional claims statements (particularly in the bulleted list in the paper) are a bit vague (see comments in PDF attached at the bottom of this review. Clear examples of the use of the tool are found in the codebase and will be sufficient to confirm this functionality when aspects like "many software auxiliaries" are made more precise.

### Performance:
The authors briefly mention that this tool is for "fast simulation" but no quantitative measure of "fast" is provided, so I have chosen to check this box at the moment. However, I would recommend that the authors either drop the "fast" qualifier or quantify it specifically.  This is just a suggestion and I'll leave it up to the authors whether to incorporate it.

### Installation Instructions:
I've checked this box, but I would recommend that the readme state more clearly that there are no FMU binaries for MacOSX, so one must build them oneself (and, in order to do so must install omc on their mac and run the fmu creation script.)  This is just a suggestion and I'll leave it up to the authors whether to incorporate it.

### Example Usage:
The repository contains examples which are clear and are discussed well in the readme. The paper could potentially benefit from a brief explanation of one of those examples. But, it's not 100% necessary. It just might bring the conceptual level of the paper discussion up to the level of discussing real world applications. This is just a suggestion and I'll leave it up to the authors whether to incorporate it.

### Quality of writing:
There are a handful of instances in which the paper is not very clear. Actually, the README was much more informative, in part because it used clearer, more direct language. I recommend the authors consider improving their paper clarity by approaching the clear direct style of the README. In any case, I have made some scribbles on [the pdf here](https://github.com/openjournals/joss-reviews/files/4963806/10.21105.joss.02435.pdf)
 which give suggestions regarding grammar, sentence structure, and diction which might improve the clarity of the work.

### References:
The references are in great need of review and effort. The list of references may not be complete. It seems as if there are some software tools heavily used which are either not cited (PyFMI) or possibly cited improperly (Open Modelica). I have made notes in [the pdf here](https://github.com/openjournals/joss-reviews/files/4963806/10.21105.joss.02435.pdf) regarding this.
All references are missing DOIs. Whedon picked up on the following, but a few more DOIs are missing beyond those. Please include all of the DOIs I have noted as missing in [the pdf](https://github.com/openjournals/joss-reviews/files/4963806/10.21105.joss.02435.pdf). Note especially the missing DOIs noticed by whedon:

- https://doi.org/10.5334/jors.188 may be missing for title: PyPSA: Python for Power System Analysis
- https://doi.org/10.1109/tpwrs.2018.2829021 may be missing for title: Pandapower: An Open-Source Python Tool for Convenient Modeling, Analysis, and Optimization of Electric Power Systems
- https://doi.org/10.1109/tie.2012.2194969 may be missing for title: Advanced Control Architectures for Intelligent Microgrids, Part I: Decentralized and Hierarchical Control
- https://doi.org/10.1109/tie.2012.2196889 may be missing for title: Advanced Control Architectures for Intelligent Microgrids, Part II: Power Quality, Energy Storage, and AC/DC Microgrids
