+++ setup.py
@@ -32,6 +32,8 @@ setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Operating System :: Microsoft :: Windows',
+        'Operating System :: POSIX :: Linux'        
     ],
     description="OpenModelica Microgrid Gym",
     install_requires=requirements,
