+++ setup.py
@@ -31,6 +31,8 @@ setup(
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Operating System :: Microsoft :: Windows',
+        'Operating System :: POSIX :: Linux'        
     ],
     description="OpenModelica Microgrid Gym",
     install_requires=requirements,
