### 2020-07-24 4d90272
Update OS support in Readme and setup.py Docstrings in Tests updated .md and .bib file according katy huffs remarks. Update paper.md
