### 2020-07-24 3b74fc3
still contructionv parts in the grid.mo, needs to be cleaned up before pushing it somewhere else
