### 2020-07-27 19162c2
changed meas setting
