### 2020-07-27 a22b706
#45 incorporated daniels feedback
