+++ openmodelica_microgrid_gym/net/net.yaml
@@ -7,7 +7,7 @@ components:
     #i_nom: 20
     #i_lim: 30
     #v_DC: 1000
-    id: ''
+    id: inverter1
     cls: MasterInverter
     in:
       u: [i1p1, i1p2, i1p3]    # names of the inputs
@@ -17,11 +17,11 @@ components:
     # iref: [0,0,0]
     # vref: [1,0,0]
   inv2:
-    id: ''
+    id: inverter2
     cls: SlaveInverter
     #pll:
-    #  p: 10
-    #  i: 200
+    #  kP: 10
+    #  kI: 200
     in:
       u: [i2p1, i2p2, i2p3]
     out:
@@ -32,17 +32,4 @@ components:
     id: rl1
     cls: Load
     out:
-      i: [inductor1.i, inductor2.i, inductor3.i]
-
-
-
-
-# model_params={'inverter1.v_DC': v_DC},
-# model_path='../fmu/grid.network.fmu',
-# model_input=['i1p1', 'i1p2', 'i1p3', 'i2p1', 'i2p2', 'i2p3'],
-# model_output=dict(lc1=[['inductor1.i', 'inductor2.i', 'inductor3.i'],
-#                        ['capacitor1.v', 'capacitor2.v', 'capacitor3.v']],
-#                   lcl1=[['inductor1.i', 'inductor2.i', 'inductor3.i'],
-#                         ['capacitor1.v', 'capacitor2.v', 'capacitor3.v'],
-#                   rl1=[f'inductor{i}.i' for i in range(1, 4)]])
-# slave sepoint: np.zeros(3)]})
+      i: [.inductor1.i, .inductor2.i, .inductor3.i]
