### 2020-07-27 b84ccb0
updated .md and .bib file according katy huffs remarks.
