+++ docs/JOSS/paper.bib
@@ -8,6 +8,7 @@
   Author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
   Title = {OpenAI Gym},
   Year = {2016},
+  url    = {https://arxiv.org/abs/1606.01540},
   Eprint = {arXiv:1606.01540},
 }
 @Misc{OSMC2020,
@@ -16,9 +17,27 @@
   year   = {2020},
   url    = {https://github.com/OpenModelica/OpenModelica},
 }
+
+
+@inproceedings{Fritzson2018,
+  author = {Peter Fritzson and Adrian Pop and Adeel Asghar and Bernhard Bachmann and Willi Braun and Robert Braun and Lena Buffoni and Francesco Casella and Rodrigo Castro and Alejandro Danós and Rüdiger Franke and Mahder Gebremedhin and Bernt Lie and Alachew Mengist and Kannan Moudgalya and Lennart Ochel and Arunkumar Palanisamy and Wladimir Schamai and Martin Sjölund and Bernhard Thiele and Volker Waurich and Per Östlund},
+  title = {{The OpenModelica Integrated Modeling, Simulation and Optimization Environment}},
+  year    = {2018},
+  crossref = {modelicaUS2018},
+  doi = {10.3384/ecp18154206}
+}
+
+@Misc{FMI2020,
+  author = {{Modelica Association}},
+  title  = {{Functional Mock-up Interface}},
+  year   = {2020},
+  url    = {https://fmi-standard.org/},
+}
+
 @Article{Kroposki2008,
   author  = {B. {Kroposki} and R. {Lasseter} and T. {Ise} and S. {Morozumi} and S. {Papathanassiou} and N. {Hatziargyriou}},
   title   = {{Making Microgrids Work}},
+  doi = {10.1109/MPE.2008.918718},
   journal = {IEEE Power and Energy Magazine},
   year    = {2008},
   volume  = {6},
@@ -28,6 +47,7 @@
 @Article{Lund2017,
   author   = {Henrik Lund and Poul Alberg {\O}stergaard and David Connolly and Brian Vad Mathiesen},
   title    = {{Smart Energy and Smart Energy Systems}},
+  doi = {10.1016/j.energy.2017.05.123},
   journal  = {Energy},
   year     = {2017},
   volume   = {137},
@@ -46,6 +66,7 @@
 @Article{Brown2018,
   author  = {T. Brown and J. H\"orsch and D. Schlachtberger},
   title   = {{PyPSA: Python for Power System Analysis}},
+  doi     = {10.5334/jors.188},
   journal = {Journal of Open Research Software},
   year    = {2018},
   volume  = {6},
@@ -55,6 +76,7 @@
 @Article{Thurner2018,
   author  = {L. {Thurner} and A. {Scheidler} and F. {Sch\"afer} and J. {Menke} and J. {Dollichon} and F. {Meier} and S. {Meinecke} and M. {Braun}},
   title   = {{Pandapower: An Open-Source Python Tool for Convenient Modeling, Analysis, and Optimization of Electric Power Systems}},
+  doi = {10.1109/TPWRS.2018.2829021},
   journal = {IEEE Transactions on Power Systems},
   year    = {2018},
   volume  = {33},
@@ -65,6 +87,7 @@
   author  = {J. M. {Guerrero} and M. {Chandorkar} and T. {Lee} and P. C. {Loh}},
   title   = {{Advanced Control Architectures for Intelligent Microgrids, Part I: Decentralized and Hierarchical Control}},
   journal = {IEEE Transactions on Industrial Electronics},
[…]
