### 2020-07-27 d3127d9
Feature/45 env norm (#56)
* First skeleton of the new network configuration file and corresponding classes
* snapshot before refactoring
* normalization and proper reference value calculation is missing. The helper functions around should be complete now.
* normalization and inverted droop for slave controller
* added normalization bool on/off to norm env
* added normalization test env from two_inv_static_droop_control
* compared results to stored results from two_inverter_static_droop_controller.py -> huge difference
* changed viz_mode from step to episode by default
* fixed the main issue wi
[…]
