### 2020-07-27 d714c54
Script from Jarren to run FPGA via SSH
