+++ docs/JOSS/paper.bib
@@ -34,6 +34,15 @@
   url    = {https://fmi-standard.org/},
 }
 
+
+@Misc{PyFMI2020,
+  author = {{Modelon AB}},
+  title  = {{PyFMI}},
+  year   = {2020},
+  url    = {https://github.com/modelon-community/PyFMI},
+}
+
+
 @Article{Kroposki2008,
   author  = {B. {Kroposki} and R. {Lasseter} and T. {Ise} and S. {Morozumi} and S. {Papathanassiou} and N. {Hatziargyriou}},
   title   = {{Making Microgrids Work}},
