+++ docs/JOSS/paper.md
@@ -34,8 +34,8 @@ bibliography: paper.bib
 
 
 Micro- and smart grids (MSG) play an important role both for integrating renewable energy sources in conventional electricity grids and for providing power supply in remote areas [@Lund2017]. 
-Due to their high efficiency and flexibility, power inverters are largely used to drive modern MSG.
-Nevertheless, controlling MSGs is a challenging task due to the high requirements on energy availability, safety and voltage quality. 
+Due to their high efficiency and flexibility, power electronic converters are largely used to drive modern MSG.
+Controlling MSGs is a challenging task due to the high requirements on energy availability, safety and voltage quality. 
 This is particularly demanding due to the wide range of different MSG topologies depending on their field of application like industrial campuses, residential areas or remote off-grid electrification [@Kroposki2008].
 This results in high demand for comprehensive testing of new control concepts during their development phase and comparisons with the state of the art to ensure their feasibility.
 This applies in particular to data-driven control approaches such as reinforcement learning (RL), the stability and operating behavior of which cannot be evaluated a priori [@Garcia2015].
@@ -47,7 +47,7 @@ Due to the component-oriented modeling framework based on OpenModelica, dynamic 
 This is an essential difference to already available open-source solutions for the simulation of electrical energy networks, which, in contrast, generally depict large-scale transmission networks with abstracted models in the (quasi)-stationary state (e.g. PyPSA [@Brown2018] or Pandapower [@Thurner2018]). Besides the pure modeling and simulation of microgrids, basic building blocks for setting up a hierarchical control framework on the inner and primary level [@Guerrero2013] are provided with ``OMG``. 
 
 
-The API is designed to provide a user-friendly interface to connect a modeled microgrid (the environment for the RL approaches) with a wide range of control methods such as classical linear feedback control or model predictive control techniques (cf. Fig. 1). Moreover, the standardized OpenAI Gym interface [@OpenAI2016] is also available for training data-driven control approaches like RL. 
+The API is designed to provide a user-friendly interface to connect a modeled microgrid (the simulation environment) with a wide range of control methods such as classical linear feedback control or model predictive control techniques (cf. Fig. 1). Moreover, the standardized OpenAI Gym interface [@OpenAI2016] is also available for training data-driven control approaches like RL. 
 Many auxiliary functionalities for the essential operation of microgrids are shipped with OMG such as coordinate transformations for basic controller classes, monitoring wrappers and phase-locked loops for frequency and phase angle extraction. 
 Following this structure, nearly every control approach including data-driven RL can be implemented and tested with ``OMG`` in a relatively short amount of time. 
 To highlight the challenges of data-driven control approaches in safety-critical environments, application examples using safe Bayesian optimization [@Berkenkamp2020] for automated controller design are provided in the toolbox. 
@@ -69,7 +69,7 @@ The user can select between a wide range of different grid components and connec
 
 * Dynamic simulation of local electricity grids on component level including single and multi-phase systems as well as AC and DC operation. 
 
-* Easy exchange of models between computing platforms and simulation of the models by using the FMI 2.0 standard with C++ code inside and PyFMI for the access in Python. Appropriate numeric solvers for the underlying system of ordinary differential equations can be easily chosen within the usual Python packages (e.g. SciPy) due to the usage of co-simulation. 
+* Easy exchange of models between computing platforms and simulation of the models by using the FMI 2.0 standard [@FMI2020] with C++ code inside and PyFMI [@FMI2020] for the access in Python. Appropriate numeric solvers for the underlying system of ordinary differential equations can be easily chosen within the usual Python packages (e.g. SciPy) due to the usage of co-simulation. 
 
[…]
