### 2020-07-28 #2435 @stheid
@katyhuff The master branch contains all changes proposed (https://github.com/upb-lea/openmodelica-microgrid-gym/commit/4d902722f090d8b0fb81b1b0c7ac5f0434338127).

right now the CI does not properly build the master, but this issue is already solved in develop. however, the develop is currently not in a mergable state, therefore i only cherry picked the changes to the pdf and readme from that branch
