+++ openmodelica_microgrid_gym/env/physical_testbench.py
@@ -1,0 +1,153 @@
+import gym
+import paramiko
+import numpy as np
+import matplotlib.pyplot as plt
+
+from openmodelica_microgrid_gym.util import dq0_to_abc
+
+
+class TestbenchEnv(gym.Env):
+
+    viz_modes = {'episode', 'step', None}
+    """Set of all valid visualisation modes"""
+
+    def __init__(self, host: str = 'lea-jde10', username: str = 'root', password: str = '',
+                 DT: float = 1/20000, executable_script_name: str = 'my_first_hps' ,num_steps: int = 1000,
+                 kP: float = 0.01, kI: float = 5.0, i_ref: float = 10.0, f_nom: float = 50.0, i_limit: float = 30,
+                 i_nominal: float = 20):
+
+        self.ssh = paramiko.SSHClient()
+        self.ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+
+        self.host = host
+        self.username = username
+        self.password = password
+        self.DT = DT
+        self.executable_script_name = executable_script_name
+        self.max_episode_steps = num_steps
+        self.kP = kP
+        self.kI = kI
+        self.i_ref = i_ref
+        self.f_nom = f_nom
+        self.data = np.array(list())
+        self.current_step = 0
+        self.done = False
+        self.i_limit = i_limit
+        self.i_nominal = i_nominal
+
+    @staticmethod
+    def __decode_result(ssh_result):
+
+        result = list()
+        for line in ssh_result.read().splitlines():
+
+            temp = line.decode("utf-8").split(",")
+
+            if len(temp) == 11:
+                temp.pop(-1)  # Drop the last item
+
+                floats = [float(i) for i in temp]
+                # print(floats)
+                result.append(floats)
+            elif len(temp) != 1:
+                print(temp)
+
+        N = (len(result))
+        decoded_result = np.array(result)
+
+        return decoded_result
+
+    def rew_fun(self, Iabc_meas, Idq0_SP, phase) -> float:
+        """
+        Defines the reward function for the environment. Uses the observations and setpoints to evaluate the quality of the
+        used parameters.
+        Takes current measurement and setpoints so calculate the mean-root-error control error and uses a logarithmic
+        barrier function in case of violating the current limit. Barrier function is adjustable using parameter mu.
+
+        :param cols: list of variable names of the data
+        :param data: observation data from the environment (ControlVariables, e.g. currents and voltages)
+        :return: Error as negative reward
+        """
+        mu = 2
+
+        # setpoints
+        Iabc_SP = dq0_to_abc(Idq0_SP, phase)  # convert dq set-points into three-phase abc coordinates
+
+        # control error = mean-root-error (MRE) of reference minus measurement
+        # (due to normalization the control error is often around zero -> compared to MSE metric, the MRE provides
+        #  better, i.e. more significant,  gradients)
+        # plus barrier penalty for violating the current constraint
+        error = np.sum((np.abs((Iabc_SP - Iabc_meas)) / self.i_limit) ** 0.5, axis=0) \
+                + -np.sum(mu * np.log(1 - np.maximum(np.abs(Iabc_meas) - self.i_nominal, 0) / \
+                (self.i_limit - self.i_nominal)), axis=0) * self.max_episode_steps
+
+        return -error.squeeze()
+
+    def reset(self, kP, kI):
+        # toDo: ssh connection not open every episode!
+        self.kP = kP
+        self.kI = kI
+
[…]
