### 2020-07-31 05553ed
hardware runner little change in comp. to runner  - controller part of env  - plotting still need to be reviewed
