+++ examples/single_inverter_current_control_safe_opt.py
@@ -18,6 +18,8 @@ from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
 from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
 from openmodelica_microgrid_gym.env import PlotTmpl
+from openmodelica_microgrid_gym.env.physical_testbench import TestbenchEnv
+from openmodelica_microgrid_gym.execution.runner_hardware import RunnerHardware
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
 
 # Choose which controller parameters should be adjusted by SafeOpt.
@@ -33,10 +35,10 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
-max_episode_steps = 1000  # number of simulation steps per episode
-num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
-v_DC = 400  # DC-link voltage / V; will be set as model parameter in the FMU
-nomFreq = 0  # nominal grid frequency / Hz
+max_episode_steps = 10000  # number of simulation steps per episode
+num_episodes = 10  # number of simulation episodes (i.e. SafeOpt iterations)
+#v_DC = 40  # DC-link voltage / V; will be set as model parameter in the FMU
+nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
 iLimit = 30  # inverter current limit / A
 iNominal = 20  # nominal inverter current / A
@@ -108,8 +110,8 @@ if __name__ == '__main__':
 
     # For 2D example, choose Kp and Ki as mutable parameters (below) and define bounds and lengthscale for both of them
     if adjust == 'Kpi':
-        bounds = [(0.0, 0.03), (0, 20)]
-        lengthscale = [.01, 5.]
+        bounds = [(0.005, 0.04), (2, 50)]
+        lengthscale = [.001, 1.]
 
     # The performance should not drop below the safe threshold, which is defined by the factor safe_threshold times
     # the initial performance: safe_threshold = 1.2 means. Performance measurement for optimization are seen as
@@ -148,7 +150,7 @@ if __name__ == '__main__':
 
     # For 2D example, choose Kp and Ki as mutable parameters
     elif adjust == 'Kpi':
-        mutable_params = dict(currentP=MutableFloat(0.01), currentI=MutableFloat(0))
+        mutable_params = dict(currentP=MutableFloat(0.01), currentI=MutableFloat(10))
         current_dqp_iparams = PI_params(kP=mutable_params['currentP'], kI=mutable_params['currentI'], limits=(-1, 1))
 
     # Define the droop parameters for the inverter of the active power Watt/Hz (DroopGain), delta_t (0.005) used for the
@@ -192,98 +194,101 @@ if __name__ == '__main__':
     # - inputs to the models are the connection points to the inverters (see user guide for more details)
     # - model outputs are the the 3 currents through the inductors and the 3 voltages across the capacitors
 
-    def xylables(fig):
-        ax = fig.gca()
-        ax.set_xlabel(r'$t\,/\,\mathrm{ms}$')
-        ax.set_ylabel('$i_{\mathrm{abc}}\,/\,\mathrm{A}$')
-        ax.grid(which='both')
-        #fig.savefig('Inductor_currents.pdf')
+    # def xylables(fig):
[…]
