+++ examples/single_inverter_current_control_safe_opt.py
@@ -36,7 +36,7 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 # Simulation definitions
 delta_t = 0.5e-4  # simulation time step size / s
 max_episode_steps = 10000  # number of simulation steps per episode
-num_episodes = 10  # number of simulation episodes (i.e. SafeOpt iterations)
+num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
 #v_DC = 40  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -255,40 +255,40 @@ if __name__ == '__main__':
 
     runner.run(num_episodes, visualise=True)
 
-    # print('\n Experiment finished with best set: \n\n {}'.format(agent.history.df[:]))
-    #
-    # print('\n Experiment finished with best set: \n')
-    # print('\n  {} = {}' .format(adjust, agent.history.df.at[np.argmax(agent.history.df['J']),'Params']))
-    # print('  Resulting in a performance of J = {}'.format(np.max(agent.history.df['J'])))
-    # print('\n\nBest experiment results are plotted in the following:')
-    #
-    #
+    print('\n Experiment finished with best set: \n\n {}'.format(agent.history.df[:]))
+
+    print('\n Experiment finished with best set: \n')
+    print('\n  {} = {}' .format(adjust, agent.history.df.at[np.argmax(agent.history.df['J']),'Params']))
+    print('  Resulting in a performance of J = {}'.format(np.max(agent.history.df['J'])))
+    print('\n\nBest experiment results are plotted in the following:')
+
+
     # # Show best episode measurment (current) plot
     # best_env_plt = runner.run_data['best_env_plt']
     # ax = best_env_plt[0].axes[0]
     # ax.set_title('Best Episode')
     # best_env_plt[0].show()
     # best_env_plt[0].savefig('best_env_plt.png')
-    #
-    # # Show last performance plot
-    # best_agent_plt = runner.run_data['last_agent_plt']
-    # ax = best_agent_plt.axes[0]
-    # ax.grid(which='both')
-    # ax.set_axisbelow(True)
-    #
-    # if adjust == 'Ki':
-    #     ax.set_xlabel(r'$K_\mathrm{i}\,/\,\mathrm{(VA^{-1}s^{-1})}$')
-    #     ax.set_ylabel(r'$J$')
-    # elif adjust == 'Kp':
-    #     ax.set_xlabel(r'$K_\mathrm{p}\,/\,\mathrm{(VA^{-1})}$')
-    #     ax.set_ylabel(r'$J$')
-    # elif adjust == 'Kpi':
-    #     agent.params.reset()
-    #     ax.set_xlabel(r'$K_\mathrm{i}\,/\,\mathrm{(VA^{-1}s^{-1})}$')
-    #     ax.set_ylabel(r'$K_\mathrm{p}\,/\,\mathrm{(VA^{-1})}$')
-    #     ax.get_figure().axes[1].set_ylabel(r'$J$')
-    #     plt.plot(bounds[0], [mutable_params['currentP'].val, mutable_params['currentP'].val], 'k-', zorder=1, lw=4,
-    #              alpha=.5)
-    # best_agent_plt.show()
-    # best_agent_plt.savefig('agent_plt.png')
+
+    # Show last performance plot
+    best_agent_plt = runner.run_data['last_agent_plt']
[…]
