### 2020-08-03 de7e23a
# 59 normalized performance measurement to inital performance
