### 2020-08-03 b1c07c0
# 59 try with kp only and render function get J as input for title
