+++ openmodelica_microgrid_gym/execution/runner_hardware.py
@@ -48,7 +48,9 @@ class RunnerHardware:
         agent_fig = None
 
         for i in tqdm(range(n_episodes), desc='episodes', unit='epoch'):
-            self.env.reset(self.agent.params[0], self.agent.params[1])
+            #self.env.reset(self.agent.params[0], self.agent.params[1])
+            self.env.reset(self.agent.params[0], 5)
+            #self.env.render(0)
             done, r = False, None
             for _ in tqdm(range(self.env.max_episode_steps), desc='steps', unit='step', leave=False):
                 self.agent.observe(r, done)
@@ -59,7 +61,7 @@ class RunnerHardware:
                 if done:
                     break
             self.agent.observe(r, done)
-            self.env.render()
+            self.env.render(self.agent.history.df.J.iloc[-1])
 
             if visualise:
                 agent_fig = self.agent.render()
