### 2020-08-03 8ab583a
#50 initial experiments
Creating Gekkofiles

Making gekko an SPM comparable

deleted fmu folder
created new experiment folder
