+++ experiments/swing_equation/microgrid_droop_ctrl.py
@@ -1,0 +1,129 @@
+#####################################
+# Example using a FMU by OpenModelica as gym environment containing two inverters, each connected via an LC-filter to
+# supply in parallel a RL load.
+# This example uses the available standard controllers as defined in the 'auxiliaries' folder.
+# One inverter is set up as voltage forming inverter with a direct droop controller.
+# The other controller is used as current sourcing inverter with an inverse droop controller which reacts on the
+# frequency and voltage change due to its droop control parameters by a power/reactive power change.
+
+import logging
+from functools import partial
+
+import matplotlib.pyplot as plt
+
+import gym
+import numpy as np
+
+from openmodelica_microgrid_gym import Runner
+from openmodelica_microgrid_gym.agents import StaticControlAgent
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQ0PIPIController, \
+    MultiPhaseDQCurrentController, InverseDroopParams, PLLParams
+
+# Simulation definitions
+delta_t = 0.5e-4  # simulation time step size / s
+max_episode_steps = 8000  # number of simulation steps per episode
+num_episodes = 1  # number of simulation episodes
+# (here, only 1 episode makes sense since simulation conditions don't change in this example)
+v_DC = 1000  # DC-link voltage / V; will be set as model parameter in the fmu
+nomFreq = 50  # grid frequency / Hz
+nomVoltPeak = 230 *1.4241 # nominal grid voltage / V
+iLimit = 30  # current limit / A
+iNominal = 20  # nominal current / A
+DroopGain = 4000.0  # virtual droop gain for active power / W/Hz
+QDroopGain = 50  # virtual droop gain for reactive power / VAR/V
+
+logging.basicConfig()
+
+
+def load_step(t, gain):
+    """
+    Defines a load step after 0.2 s
+    Doubles the load parameters
+    :param t:
+    :param gain: device parameter
+    :return: Dictionary with load parameters
+    """
+    return 1*gain if t < .52 else 2*gain
+
+
+if __name__ == '__main__':
+    ctrl = []  # Empty dict which shall include all controllers
+
+    #####################################
+    # Define the voltage forming inverter as master
+    # Voltage control PI gain parameters for the voltage sourcing inverter
+   # voltage_dqp_iparams = PI_params(kP=0.025, kI=60, limits=(-iLimit, iLimit))
+    # Current control PI gain parameters for the voltage sourcing inverter
+    #current_dqp_iparams = PI_params(kP=0.012, kI=90, limits=(-1, 1))
+    #pll_params = PLLParams(kP=20, kI=400, limits=None, f_nom=nomFreq)
+
+    current_dqp_iparams = PI_params(kP=0.005, kI=200, limits=(-1, 1))
+    # PI gain parameters for the PLL in the current forming inverter
+    pll_params = PLLParams(kP=10, kI=200, limits=None, f_nom=nomFreq)
+    # Droop characteristic for the active power Watt/Hz, delta_t
+    droop_param = InverseDroopParams(DroopGain, delta_t, nomFreq, tau_filt=0.04)
+    # Droop characteristic for the reactive power VAR/Volt Var.s/Volt
+    qdroop_param = InverseDroopParams(QDroopGain, delta_t, nomVoltPeak, tau_filt=0.01)
+    # Add to dict
+    ctrl.append(MultiPhaseDQCurrentController(current_dqp_iparams, pll_params, delta_t, iLimit,
+                                              droop_param, qdroop_param, name='slave1'))
+
+    #####################################
+    # Define the current sourcing inverter as slave
[…]
