### 2020-08-05 b831993
# 50 more values (Setpoints) back from FPGA
