### 2020-08-05 a423173
# 50 - give setpoints in abc from FPGA to reward function - add try except to ssh connection - plot setpoints
