+++ openmodelica_microgrid_gym/env/physical_testbench.py
@@ -1,8 +1,12 @@
+from socket import socket
+
 import gym
 import paramiko
 import numpy as np
 import matplotlib.pyplot as plt
-from time import strftime, gmtime
+from time import strftime, gmtime, sleep
+
+from paramiko import BadHostKeyException, AuthenticationException, SSHException
 
 from openmodelica_microgrid_gym.util import dq0_to_abc
 
@@ -12,7 +16,7 @@ class TestbenchEnv(gym.Env):
     viz_modes = {'episode', 'step', None}
     """Set of all valid visualisation modes"""
 
-    def __init__(self, host: str = 'lea-jde10', username: str = 'root', password: str = '',
+    def __init__(self, host: str = '131.234.172.139', username: str = 'root', password: str = '',
                  DT: float = 1/20000, executable_script_name: str = 'my_first_hps' ,num_steps: int = 1000,
                  kP: float = 0.01, kI: float = 5.0, i_ref: float = 10.0, f_nom: float = 50.0, i_limit: float = 30,
                  i_nominal: float = 20):
@@ -44,7 +48,7 @@ class TestbenchEnv(gym.Env):
 
             temp = line.decode("utf-8").split(",")
 
-            if len(temp) == 11:
+            if len(temp) == 20:
                 temp.pop(-1)  # Drop the last item
 
                 floats = [float(i) for i in temp]
@@ -58,7 +62,7 @@ class TestbenchEnv(gym.Env):
 
         return decoded_result
 
-    def rew_fun(self, Iabc_meas, Idq0_SP, phase) -> float:
+    def rew_fun(self, Iabc_meas, Iabc_SP) -> float:
         """
         Defines the reward function for the environment. Uses the observations and setpoints to evaluate the quality of the
         used parameters.
@@ -72,7 +76,7 @@ class TestbenchEnv(gym.Env):
         mu = 2
 
         # setpoints
-        Iabc_SP = dq0_to_abc(Idq0_SP, phase)  # convert dq set-points into three-phase abc coordinates
+        #Iabc_SP = dq0_to_abc(Idq0_SP, phase)  # convert dq set-points into three-phase abc coordinates
 
         # control error = mean-root-error (MRE) of reference minus measurement
         # (due to normalization the control error is often around zero -> compared to MSE metric, the MRE provides
@@ -89,7 +93,28 @@ class TestbenchEnv(gym.Env):
         self.kP = kP
         self.kI = kI
 
-        self.ssh.connect(self.host, username=self.username, password=self.password)
+        #self.ssh.connect(self.host, username=self.username, password=self.password)
+
+        count_retries = 0
+        connected = False
+
+        #for x in range(10):
+        while not connected and count_retries < 10:
+            count_retries+=1
+            try:
+                print('I Try!')
+                self.ssh.connect(self.host, username=self.username, password=self.password)
+                connected =  True
+                #return True
+            except (BadHostKeyException, AuthenticationException,
+                    SSHException, socket.gaierror) as e:
+                print(e)
+                print('Argh')
+                sleep(1)
+
+        if count_retries == 10:
+            print( 'SSH FUCKED UP!')
+
 
 
         #toDo: get SP and kP/I from agent?
@@ -113,10 +138,11 @@ class TestbenchEnv(gym.Env):
[…]
