### 2020-08-07 368a21a
# 59 - store measurement data  - ssh connect catch all exceptions
