### 2020-08-10 f41381a
# 50 - controller layout using mar.opt - loop for find max lengthscale (check and store if unsafe)
