+++ examples/single_inverter_current_control_safe_opt.py
@@ -5,6 +5,7 @@
 
 
 import logging
+from time import strftime, gmtime
 from typing import List
 
 import GPy
@@ -35,13 +36,16 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
 
 
 include_simulate = True
+do_measurement = False
 
+lengthscale_vec = np.linspace(0.05,2,20)
+unsafe_vec = np.zeros(20)
 
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
 max_episode_steps = 1000  # number of simulation steps per episode
-num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
-#v_DC = 40  # DC-link voltage / V; will be set as model parameter in the FMU
+num_episodes = 100  # number of simulation episodes (i.e. SafeOpt iterations)
+v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
 iLimit = 30  # inverter current limit / A
@@ -50,6 +54,20 @@ mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
 i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+
+# Controller layout due to magniitude optimum:
+L = 2.2e-3  # / H
+R = 585e-3  # / Ohm
+tau_plant = L/R
+gain_plant = 1/R
+
+# take inverter into account uning s&h (exp(-s*delta_T/2))
+
+Tn = tau_plant   # Due to compensate
+Kp_init = tau_plant/(2*delta_t*gain_plant*v_DC)
+Ki_init = Kp_init/(Tn)
+
+
 
 
 class Reward:
@@ -95,219 +113,269 @@ class Reward:
 
 
 if __name__ == '__main__':
-    #####################################
-    # Definitions for the GP
-    prior_mean = 0#2  # mean factor of the GP prior mean which is multiplied with the first performance of the initial set
-    noise_var = 0.001#0.001 ** 2  # measurement noise sigma_omega
-    prior_var = 2#0.1  # prior variance of the GP
 
-    bounds = None
-    lengthscale = None
-    if adjust == 'Kp':
-        bounds = [(0, 10)]  # bounds on the input variable Kp
-        lengthscale = [.75]  # length scale for the parameter variation [Kp] for the GP
+    for ll in range(len(lengthscale_vec)):
+        #####################################
+        # Definitions for the GP
+        prior_mean = 0#2  # mean factor of the GP prior mean which is multiplied with the first performance of the initial set
+        noise_var = 0.001#0.001 ** 2  # measurement noise sigma_omega
+        prior_var = 2#0.1  # prior variance of the GP
 
-    # For 1D example, if Ki should be adjusted
-    if adjust == 'Ki':
-        bounds = [(0, 30)]  # bounds on the input variable Ki
-        lengthscale = [5.]  # length scale for the parameter variation [Ki] for the GP
+        bounds = None
+        lengthscale = None
+        if adjust == 'Kp':
+            bounds = [(0, 10)]  # bounds on the input variable Kp
+            lengthscale = [.75]  # length scale for the parameter variation [Kp] for the GP
 
-    # For 2D example, choose Kp and Ki as mutable parameters (below) and define bounds and lengthscale for both of them
-    if adjust == 'Kpi':
-        bounds = [(0.0, 8), (0, 100)]
-        lengthscale = [2., 20.]
+        # For 1D example, if Ki should be adjusted
[…]
