### 2020-08-10 a418055
# 50 - store worst plot, too
