### 2020-08-10 17cd692
# 50 - store worst plot, too
