+++ examples/measure.py
@@ -1,0 +1,247 @@
+#####################################
+# Example using a FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
+# Simulation setup: Single inverter supplying 15 A d-current to an RL-load via a LC filter
+# Controller: PI current controller gain parameters are optimized by SafeOpt
+
+
+import logging
+from time import strftime, gmtime
+from typing import List
+
+import GPy
+import gym
+import numpy as np
+
+import matplotlib.pyplot as plt
+
+from openmodelica_microgrid_gym import Runner
+from openmodelica_microgrid_gym.agents import SafeOptAgent
+from openmodelica_microgrid_gym.agents.util import MutableFloat
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.env import PlotTmpl
+from openmodelica_microgrid_gym.env.physical_testbench import TestbenchEnv
+from openmodelica_microgrid_gym.execution.runner_hardware import RunnerHardware
+from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
+
+# Choose which controller parameters should be adjusted by SafeOpt.
+# - Kp: 1D example: Only the proportional gain Kp of the PI controller is adjusted
+# - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
+# - Kpi: 2D example: Kp and Ki are adjusted simultaneously
+
+adjust = 'Kpi'
+
+# Check if really only one simulation scenario was selected
+if adjust not in {'Kp', 'Ki', 'Kpi'}:
+    raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
+
+unsafe_vec = np.zeros(20)
+
+# Simulation definitions
+delta_t = 1e-4  # simulation time step size / s
+max_episode_steps = 1000  # number of simulation steps per episode
+num_episodes = 200  # number of simulation episodes (i.e. SafeOpt iterations)
+v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
+nomFreq = 50  # nominal grid frequency / Hz
+nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
+iLimit = 30  # inverter current limit / A
+iNominal = 20  # nominal inverter current / A
+mu = 2  # factor for barrier function (see below)
+DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
+QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
+i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+
+# Controller layout due to magniitude optimum:
+L = 2.2e-3  # / H
+R = 585e-3  # / Ohm
+tau_plant = L / R
+gain_plant = 1 / R
+
+# take inverter into account uning s&h (exp(-s*delta_T/2))
+
+Tn = tau_plant  # Due to compensate
+Kp_init = tau_plant / (2 * delta_t * gain_plant * v_DC)
+Ki_init = Kp_init / (Tn)
+
+
+class Reward:
+    def __init__(self):
+        self._idx = None
+
+    def set_idx(self, obs):
+        if self._idx is None:
+            self._idx = nested_map(
+                lambda n: obs.index(n),
+                [[f'rl.inductor{k}.i' for k in '123'], 'master.phase', [f'master.SPI{k}' for k in 'dq0']])
+
+    def rew_fun(self, cols: List[str], data: np.ndarray) -> float:
+        """
+        Defines the reward function for the environment. Uses the observations and setpoints to evaluate the quality of the
[…]
