+++ examples/plotReward.py
@@ -13,11 +13,12 @@ if __name__ == '__main__':
     kP = rew[0][1]
     kI = rew[0][2]
     reward = rew[0][0]
+    # reward[0][0] = -1616 # correktion for sim - initial kp/i have to be adjusted
     xx, yy = np.meshgrid(kI, kP, sparse=True)
 
 
     #alles größer 500 auf 500 setzen??
-    reward[reward>390] = 390
+    reward[reward<-550] = -500
 
     fig = plt.figure()
     ax = fig.gca(projection='3d')
@@ -27,7 +28,7 @@ if __name__ == '__main__':
     # ax.set_zlim(-1.01, 1.01)
     #ax.zaxis.set_major_locator(LinearLocator(10))
     #ax.zaxis.set_major_formatter(FormatStrFormatter('%.02f'))
-    ax.view_init(30, 150)
+    ax.view_init(50, 30)
     ax.set_xlabel(r'$K_\mathrm{i}\,/\,\mathrm{(A^{-1}s^{-1})}$')
     ax.set_ylabel(r'$K_\mathrm{p}\,/\,\mathrm{(A^{-1})}$')
     ax.set_zlabel(r'$i_\mathrm{RME+Barrier}$')
