### 2020-08-14 #2435 @gonsie
## gonsie review

Sorry for the delay in my review. Unfortunately, I had many technical issues: I only have access to a Mac and I seem unable to mark the checkboxes in my review section at the top of this issue 😞 That said, here is my best effort at reviewing the documentation and paper.

- [X] conflict of interest
- [X] code of conduct
- [X] repo
- [X] License
- [X] contribution and authorship
- [X] installation
  Does not install on Mac. Unclear why this is not possible, nothing in PyFMI documentation indicates that Mac is an issue, yet pip install fails. Readme does make a note of this.
- [ ] functionality **unable to evaluate**
- [ ] performance **unable to evaluate**
- [ ] statement of need
  Does not exist within documentation
  Make a note of the issues with Mac on documentation page.
- [ ] example usage
  The examples seem incomplete (though I may be missing some context). The pages under Basic Examples are incomplete and missing instructions/details. The ‘creating an environment' page does not have any instructions. The image on 'creating an agent...' page does not load correctly.
- [X] functionality documentation
- [X] automated tests
- [X] community guidelines
  It would be great if the CONTRIBUTING guide could be found on the website.
- [ ] summary
  The paper starts with the statement of need and never quite provides a clear summary of the contribution of this software. In one paragraph, please identify the key contributions of this software.
- [X] statement of need
- [ ] state of the field
  It would be great if the features section clearly identified which aspects are unique contributions of this work. Point 3 mentions "easy exchange of models between computing platforms", citing PyFMI. Yet, PyFMI is a limiting factor when building this software for Mac. Does OMG contribute to the portability, or merely take advantage of existing (or not) portability of its dependencies?
  This features section also mentions that OMG many examples, use cases, and software auxiliaries. Please identify these (only examples are made explicit) or cite your arxiv whitepaper.
  The 'state of the field' section is missing any comparison with similar work.
- [ ] quality of writing
  Needs some grammatical edits. Both "i.e." and "e.g." should be followed by a comma (e.g., like so).
- [ ] references
  Some references in the area machine learning / reinforcement learning are needed.
