### 2020-08-17 093c21e
# 59 added noise measurement
