+++ examples/single_inverter_current_control_safe_opt.py
@@ -35,15 +35,17 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
 
-include_simulate = False
-do_measurement = True
+include_simulate = True
+do_measurement = False
+
+safe_results = False
 
 lengthscale_vec = np.linspace(0.5,0.5,1)
 unsafe_vec = np.zeros(20)
 
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
-max_episode_steps = 10000  # number of simulation steps per episode
+max_episode_steps = 1000  # number of simulation steps per episode
 num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
@@ -53,10 +55,10 @@ iNominal = 20  # nominal inverter current / A
 mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
-i_ref = np.array([0, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
-i_noise = 0.1 # Current measurement noise detected from testbench
+i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+i_noise = 0.11 # Current measurement noise detected from testbench
 
-# Controller layout due to magniitude optimum:
+# Controller layout due to magnitude optimum:
 L = 2.2e-3  # / H
 R = 585e-3  # / Ohm
 tau_plant = L/R
@@ -275,7 +277,8 @@ if __name__ == '__main__':
                            model_output=dict(rl=[['inductor1.i', 'inductor2.i', 'inductor3.i']],
                                              #inverter1=['inductor1.i', 'inductor2.i', 'inductor3.i']
                                              ),
-                           history=FullHistory()
+                           history=FullHistory(),
+                           measurement_noise = i_noise
                            )
 
             runner = Runner(agent, env)
@@ -291,8 +294,8 @@ if __name__ == '__main__':
             #####################################
             # Performance results and parameters as well as plots are stored in folder pipi_signleInv
             #agent.history.df.to_csv('len_search/result.csv')
-
-            env.history.df.to_pickle('Simulation')
+            if safe_results:
+                env.history.df.to_pickle('Simulation')
 
             # Show best episode measurment (current) plot
             best_env_plt = runner.run_data['best_env_plt']
@@ -328,7 +331,8 @@ if __name__ == '__main__':
                 plt.plot(bounds[0], [mutable_params['currentP'].val, mutable_params['currentP'].val], 'k-', zorder=1, lw=4,
                          alpha=.5)
             best_agent_plt.show()
-            best_agent_plt.savefig('len_search/agent_plt.png')
+            if safe_results:
+                best_agent_plt.savefig('len_search/agent_plt.png')
 
         if do_measurement:
             #####################################
@@ -355,8 +359,8 @@ if __name__ == '__main__':
             # ax.set_title('Best Episode')
             # best_env_plt[0].show()
             # best_env_plt[0].savefig('best_env_plt.png')
-
-            agent.history.df.to_csv('hardwareTest_plt/result.csv')
+            if safe_results:
+                agent.history.df.to_csv('hardwareTest_plt/result.csv')
 
             # Show last performance plot
[…]
