### 2020-08-17 b06c5ec
# 59 added noise measurement
