### 2020-08-17 b4291c1
# 59 added noise measurement using mean+std
