### 2020-08-17 b8ae140
# 59 lengthscale sweep for finding optimal for measurement
