### 2020-08-17 #2435 @stheid
@gonsie Thank you for your review. We will discuss it this friday in our team and try to also find a solution for mac, but it will probably be impossible to provide lasting OSX support as our group does not have regular access to a Mac.

In general it is possible to use FMI on Mac, but the precompiled FMUs provided by us are only built for Linux and Windows as we have not mac available and didn't deem the work in crosscompilation reasonable.
