### 2020-08-17 #2435 @gonsie
Thanks @stheid. Perhaps a container image could be made available?

In any case, I’m sure that @katyhuff ‘s review covers most of the issues with actually running the software.
