+++ examples/single_inverter_current_control_safe_opt.py
@@ -35,8 +35,8 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
 
-include_simulate = True
-do_measurement = False
+include_simulate = False
+do_measurement = True
 
 safe_results = False
 
@@ -45,7 +45,7 @@ unsafe_vec = np.zeros(20)
 
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
-max_episode_steps = 1000  # number of simulation steps per episode
+max_episode_steps = 100000  # number of simulation steps per episode
 num_episodes = 1  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
@@ -55,7 +55,7 @@ iNominal = 20  # nominal inverter current / A
 mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
-i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+i_ref = np.array([0, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
 #i_noise = 0.11 # Current measurement noise detected from testbench
 i_noise = np.array([[0.2, 0.073], [0.4, 0.108], [0.093, 0.109]])
 
