### 2020-08-19 abf62f1
# 59 fixed noise measurement
