+++ openmodelica_microgrid_gym/env/physical_testbench.py
@@ -199,7 +199,7 @@ class TestbenchEnv(gym.Env):
                            'm_0': self.data[:, 18]})
 
         #df.to_pickle('Measurement')
-        df.to_pickle('Noise_measurement2')
+        df.to_pickle('Noise_measurement')
 
         #plt.plot(t, V_A, t, V_B, t, V_C)
         #plt.ylabel('Voltages (V)')
