### 2020-08-19 747cec2
# 59 fixed noise measurement
