### 2020-08-19 d1d7f67
# 59 fixed noise mean
