### 2020-08-20 9dc1afe
# 59 - Current controller layout working - votlage controller started. first measurements
