+++ openmodelica_microgrid_gym/agents/safeopt.py
@@ -190,8 +190,8 @@ class SafeOptAgent(StaticControlAgent):
         else:
             logger.warning('Choose appropriate numer of control parameters')
 
-        #plt.show()             # only comment for lengthscale sweep
-        plt.close(figure)       # only needed for lengthscale sweep
+        plt.show()             # only comment for lengthscale sweep
+        #plt.close(figure)       # only needed for lengthscale sweep
         return figure
 
     def prepare_episode(self):
