+++ examples/testbenchCurrent_voltageController.py
@@ -17,7 +17,8 @@ import matplotlib.pyplot as plt
 from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
-from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController, \
+    MultiPhaseDQ0PIPIController
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.env.physical_testbench import TestbenchEnv
 from openmodelica_microgrid_gym.execution.runner_hardware import RunnerHardware
@@ -46,7 +47,7 @@ unsafe_vec = np.zeros(20)
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
 max_episode_steps = 1000  # number of simulation steps per episode
-num_episodes = 200  # number of simulation episodes (i.e. SafeOpt iterations)
+num_episodes = 100  # number of simulation episodes (i.e. SafeOpt iterations)
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -55,7 +56,7 @@ iNominal = 20  # nominal inverter current / A
 mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
-i_ref = np.array([5, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
+ref = np.array([20, 0, 0])  # exemplary set point i.e. id = 30, iq = 0, i0 = 0 / V
 #i_noise = 0.11 # Current measurement noise detected from testbench
 i_noise = np.array([[0.0, 0.0822], [0.0, 0.103], [0.0, 0.136]])
 
@@ -178,8 +179,13 @@ if __name__ == '__main__':
 
         # For 2D example, choose Kp and Ki as mutable parameters
         elif adjust == 'Kpi':
-            mutable_params = dict(currentP=MutableFloat(Kp_init), currentI=MutableFloat(Ki_init))
-            current_dqp_iparams = PI_params(kP=mutable_params['currentP'], kI=mutable_params['currentI'], limits=(-1, 1))
+            #mutable_params = dict(currentP=MutableFloat(Kp_init), currentI=MutableFloat(Ki_init))
+
+            mutable_params = dict(voltageP=MutableFloat(0.1),
+                                  voltageI=MutableFloat(10))
+            current_dqp_iparams = PI_params(kP=Kp_init, kI=Ki_init, limits=(-1, 1))
+            voltage_dqp_iparams = PI_params(kP=mutable_params['voltageP'], kI=mutable_params['voltageI'],
+                                            limits=(-iLimit, iLimit))
 
         # Define the droop parameters for the inverter of the active power Watt/Hz (DroopGain), delta_t (0.005) used for the
         # filter and the nominal frequency
@@ -191,8 +197,8 @@ if __name__ == '__main__':
         qdroop_param = DroopParams(QDroopGain, 0.002, nomVoltPeak)
 
         # Define a current sourcing inverter as master inverter using the pi and droop parameters from above
-        ctrl = MultiPhaseDQCurrentSourcingController(current_dqp_iparams, delta_t, droop_param, qdroop_param,
-                                                     undersampling=1, name='master')
[…]
