### 2020-08-21 a05aa98
# 59 - tests Sim - added zero noise if none is defined for env
