### 2020-08-21 732e75a
# 59 - added SPs to agent plots
