### 2020-08-21 ff67031
#58 should be solved. added a wrapper arround pyfmi
