+++ openmodelica_microgrid_gym/env/modelica.py
@@ -11,11 +11,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 import scipy
 from matplotlib.figure import Figure
-from pyfmi import load_fmu
-from pyfmi.fmi import FMUModelME2
 from scipy import integrate
 
 from openmodelica_microgrid_gym.env.plot import PlotTmpl
+from openmodelica_microgrid_gym.env.pyfmi import PyFMI_Wrapper
 from openmodelica_microgrid_gym.net.net import Network
 from openmodelica_microgrid_gym.util import FullHistory, EmptyHistory
 
@@ -101,11 +100,7 @@ class ModelicaEnv(gym.Env):
         self.solver_method = solver_method
 
         # load model from fmu
-        model_name = basename(model_path)
-        logger.debug("Loading model {}".format(model_name))
-        self.model: FMUModelME2 = load_fmu(model_path,
-                                           log_file_name=datetime.now().strftime(f'%Y-%m-%d_{model_name}.txt'))
-        logger.debug("Successfully loaded model {}".format(model_name))
+        self.model = PyFMI_Wrapper.load(model_path)
 
         # if you reward policy is different from just reward/penalty - implement custom step method
         self.reward = reward_fun
@@ -166,28 +161,6 @@ class ModelicaEnv(gym.Env):
         self.action_space = gym.spaces.Box(low=np.full(d_i, -1), high=np.full(d_i, 1))
         self.observation_space = gym.spaces.Box(low=np.full(d_o, -np.inf), high=np.full(d_o, np.inf))
 
-    def _setup_fmu(self):
-        """
-        Initialize fmu model in self.model
-        """
-
-        self.model.setup_experiment(start_time=self.time_start)
-        self.model.enter_initialization_mode()
-        self.model.exit_initialization_mode()
-
-        e_info = self.model.get_event_info()
-        e_info.newDiscreteStatesNeeded = True
-        # Event iteration
-        while e_info.newDiscreteStatesNeeded:
-            self.model.enter_event_mode()
-            self.model.event_update()
-            e_info = self.model.get_event_info()
-
-        self.model.enter_continuous_time_mode()
-
-        # precalculating indices for more efficient lookup
-        self.model_output_idx = np.array([self.model.get_variable_valueref(k) for k in self.model_output_names])
-
     def _calc_jac(self, t, x) -> np.ndarray:  # noqa
         """
         Compose Jacobian matrix from the directional derivatives of the FMU model.
@@ -198,13 +171,7 @@ class ModelicaEnv(gym.Env):
         :param x: state (ignored)
         :return: the Jacobian matrix
         """
-        # get state and derivative value reference lists
-        refs = [[s.value_reference for s in getattr(self.model, attr)().values()]
-                for attr in
-                ['get_states_list', 'get_derivatives_list']]
-        jacobian = np.identity(len(refs[1]))
-        np.apply_along_axis(lambda col: self.model.get_directional_derivative(*refs, col), 0, jacobian)
-        return jacobian
+        return self.model.jacc()
 
     def _get_deriv(self, t: float, x: np.ndarray) -> np.ndarray:
         """
@@ -215,10 +182,10 @@ class ModelicaEnv(gym.Env):
         :return: 1d float array of derivatives
         """
         self.model.time = t
-        self.model.continuous_states = x.copy(order='C')
+        self.model.states = x.copy(order='C')
 
         # Compute the derivative
-        dx = self.model.get_derivatives()
+        dx = self.model.deriv
         return dx
 
     def _simulate(self) -> np.ndarray:
@@ -231,15 +198,15 @@ class ModelicaEnv(gym.Env):
         logger.debug(f'Simulation started for time interval {self.sim_time_interval[0]}-{self.sim_time_interval[1]}')
 
         # Advance
[…]
