### 2020-08-21 ffabc71
Updating Readme for Mac
