+++ docs/JOSS/paper.md
@@ -78,8 +78,8 @@ The user can select between a wide range of different grid components and connec
 
 
 # Availability and implementation
-``OMG`` is supported and tested on Linux and Windows. The package can be 
-installed by `pip` Python package manager using 
+``OMG`` is supported and tested on Linux and Windows. Mac users are asked to run this toolbox on a Linux VM. 
+The package should be installed in a conda environment. ``PyFMI`` can be installed via `conda install -c conda-forge pyfmi`, the package by `pip` Python package manager using 
 `pip install openmodelica_microgrid_gym` command. The source code, guide and 
 datasets are available on the GitHub repository (https://github.com/upb-lea/openmodelica-microgrid-gym). 
 
