### 2020-08-21 a018bf5
Updating JOSS for Mac, added information to PyFMI since Gonsie hat problems with the installations, might have been reasoned by the pip and PyFMI issue
