### 2020-08-21 2c6f5e0
Update paper.md
