### 2020-08-24 4a6c262
# 59 - set initial parameters e.g. R,L,... using self.self.model_parameters ind ev.reset()
