### 2020-08-24 202ebd1
# 59 - generated load class to draw from normal distribution - set random generator - able to take balanced or unbalanced loads
