### 2020-08-24 51d46aa
# 59 - shifted initial parameter setting from modelica.py to pyfmi.py reset
