+++ openmodelica_microgrid_gym/agents/episodic.py
@@ -1,0 +1,14 @@
+from openmodelica_microgrid_gym import Agent
+
+
+class EpisodicLearnerAgent(Agent):
+
+    def __init__(self):
+        pass
+
+    def update_params(self):
+        pass
+
+    @property
+    def performance(self):
+        pass
