### 2020-08-25 5ffdcbe
#60 - added flexible folder in save command
