### 2020-08-25 2435d5b
#60 - added flexible folder in save command (#71)
Co-authored-by: Daniel Weber <weber@lea.uni-paderborn.de>
