### 2020-08-25 40e2f66
#59 - include episodic agent to safeopt agent Change performance to property (getter and setter)
