+++ openmodelica_microgrid_gym/agents/episodic.py
@@ -1,14 +1,20 @@
-from openmodelica_microgrid_gym import Agent
+from openmodelica_microgrid_gym.agents import Agent
 
 
 class EpisodicLearnerAgent(Agent):
 
-    def __init__(self):
-        pass
+    def observe(self, reward: float, terminated: bool):
+        """
+        Observes current reward.
+        The idea of the episodic learner is to only learn after termination, therefore update_params() must not be
+        called during the episode but only at the end.
+        """
+        if terminated:
+            self.update_params()
 
     def update_params(self):
         pass
 
     @property
     def performance(self):
-        pass
+        return None
