+++ examples/single_inverter_current_control_safe_opt.py
@@ -16,7 +16,6 @@ import numpy as np
 
 import matplotlib.pyplot as plt
 
-from openmodelica_microgrid_gym import Runner
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
 from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
@@ -37,7 +36,6 @@ adjust = 'Kpi'
 if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
-
 include_simulate = True
 do_measurement = False
 
@@ -49,7 +47,7 @@ current_directory = os.getcwd()
 save_folder = os.path.join(current_directory, r'test_save')
 os.makedirs(save_folder, exist_ok=True)
 
-lengthscale_vec = np.linspace(0.5,0.5,1)
+lengthscale_vec = np.linspace(0.5, 0.5, 1)
 unsafe_vec = np.zeros(20)
 
 np.random.seed(0)
@@ -58,8 +56,8 @@ np.random.seed(0)
 delta_t = 1e-4  # simulation time step size / s
 max_episode_steps = 1000  # number of simulation steps per episode
 num_episodes = 2  # number of simulation episodes (i.e. SafeOpt iterations)
-n_MC = 3    # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
-            # distribution to represent real world more accurate
+n_MC = 3  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
+# distribution to represent real world more accurate
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
 nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
@@ -69,38 +67,46 @@ mu = 2  # factor for barrier function (see below)
 DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
 i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
-#i_noise = 0.11 # Current measurement noise detected from testbench
+# i_noise = 0.11 # Current measurement noise detected from testbench
 i_noise = np.array([[0.0, 0.0822], [0.0, 0.103], [0.0, 0.136]])
 
 # Controller layout due to magnitude optimum:
 L = 2.2e-3  # / H
-R = 170e-3#585e-3  # / Ohm
-tau_plant = L/R
-gain_plant = 1/R
+R = 170e-3  # 585e-3  # / Ohm
+tau_plant = L / R
+gain_plant = 1 / R
 
 # take inverter into account using s&h (exp(-s*delta_T/2))
-Tn = tau_plant   # Due to compensate
-Kp_init = tau_plant/(2*delta_t*gain_plant*v_DC)
-Ki_init = Kp_init/(Tn)
+Tn = tau_plant  # Due to compensate
+Kp_init = tau_plant / (2 * delta_t * gain_plant * v_DC)
+Ki_init = Kp_init / Tn
 
-#model_params={'rl.resistor1.R': partial(load_step, gain=R)}
-#model_params={'rl.resistor1.R': load(R)}
-#model_params={'rl.inductor1.L': load(L).load_step}
 
 class Load:
 
-    def __init__(self, load_mean, load_std = None, balanced = True):
+    def __init__(self, load_mean: float, load_std: float = None, balanced: bool = True):
+        """
[…]
