### 2020-08-25 c952e19
#68 rename has_worsend and add performance setter
