### 2020-08-25 4634856
#59 - Warning refactoring
