### 2020-08-26 e9a8874
Fixing broken paths in examples (#73)
