+++ docs/parts/user_guide/examples/basic_agent.rst
@@ -5,7 +5,7 @@ Additionally to the environment, an an agent will be created and a runner will b
 of agents and environments, as well as the execution of multiple episodes. The class will handle all information
 exchange between agent and environment like presented in the high level code architecture shown below:
 
-.. figure:: ../../pictures/highlevel.png
+.. figure:: ../../../pictures/highlevel.png
    :width: 400
    :alt:
 
@@ -13,5 +13,5 @@ Since the inputs are used for both the agent and the environment, they are defin
 
 The environment is the same as above. Afterwards, the agent and the runner get defined, and the runner runs for one episode.
 
-.. literalinclude:: ../../../examples/simple_agent.py
+.. literalinclude:: ../../../../examples/simple_agent.py
    :linenos:
