### 2020-08-26 bf8c385
#58 - hopefully fixed using model.initialize it is possible to change parameters and get model outputs
