### 2020-08-28 ae75ca8
Hotfixes for updating dev in the master:
- Updating the initialisation process in the modelica.py and pyfmi.py to ba able to set the params
- Including the datapathissue windows/linux from #74
- Commenting the creation of the "4D"-plot in the runner.py
