### 2020-08-28 78702c6
#68 - added stochastic components - shifted plot.show/close to callback
