+++ examples/single_inverter_current_control_safe_opt.py
@@ -16,11 +16,13 @@ import numpy as np
 
 import matplotlib.pyplot as plt
 
+from openmodelica_microgrid_gym import Agent
 from openmodelica_microgrid_gym.agents import SafeOptAgent
 from openmodelica_microgrid_gym.agents.util import MutableFloat
 from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
 from openmodelica_microgrid_gym.env import PlotTmpl
 from openmodelica_microgrid_gym.env.physical_testbench import TestbenchEnv
+from openmodelica_microgrid_gym.env.stochastic_components import Load, Noise
 from openmodelica_microgrid_gym.execution.monte_carlo_runner import MonteCarloRunner
 from openmodelica_microgrid_gym.execution.runner_hardware import RunnerHardware
 from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
@@ -37,6 +39,7 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
 include_simulate = True
+show_plots = True
 do_measurement = False
 
 # If True: Results are stored to directory mentioned in: REBASE to DEV after MERGE #60!!
@@ -44,7 +47,7 @@ safe_results = False
 
 # Files saves results and  resulting plots to the folder saves_VI_control_safeopt in the current directory
 current_directory = os.getcwd()
-save_folder = os.path.join(current_directory, r'test_save')
+save_folder = os.path.join(current_directory, r'MCTest_bal')
 os.makedirs(save_folder, exist_ok=True)
 
 lengthscale_vec = np.linspace(0.5, 0.5, 1)
@@ -68,11 +71,12 @@ DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
 QDroopGain = 1000.0  # virtual droop gain for reactive power / VAR/V
 i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0 / A
 # i_noise = 0.11 # Current measurement noise detected from testbench
-i_noise = np.array([[0.0, 0.0822], [0.0, 0.103], [0.0, 0.136]])
+# i_noise = np.array([[0.0, 0.0822], [0.0, 0.103], [0.0, 0.136]])
 
 # Controller layout due to magnitude optimum:
-L = 2.2e-3  # / H
-R = 170e-3  # 585e-3  # / Ohm
+L = 2.3e-3  # / H
+R = 400e-3#170e-3  # 585e-3  # / Ohm
+#R = 585e-3#170e-3  # 585e-3  # / Ohm
 tau_plant = L / R
 gain_plant = 1 / R
 
@@ -81,45 +85,15 @@ Tn = tau_plant  # Due to compensate
 Kp_init = tau_plant / (2 * delta_t * gain_plant * v_DC)
 Ki_init = Kp_init / Tn
 
+#Kp_init = 0
 
-class Load:
 
-    def __init__(self, load_mean: float, load_std: float = None, balanced: bool = True):
-        """
-        Load class which defines stochastic load. Samples load value from normal Gaussian distribution (GD) using given
-        mean and standard deviation
-        :param load_mean: Mean of the GD the load is sampled from
-        :param load_std: Standard deviation of the GD the load is sampled from
-        :param balanced: If True: all 3 phases are takes as equal; if False symmetrical load is applied
-        """
+class pltManager:
 
-        self.balanced = balanced
-        self.load_mean = load_mean
-        if load_std is None:
-            self.load_std = self.load_mean * 0.1
-        else:
[…]
