### 2020-08-28 3beb9ea
#68 inculded seed, set performance correctly
