### 2020-08-28 6e12e07
Fixing issue #77
