+++ openmodelica_microgrid_gym/env/pyfmi.py
@@ -23,18 +23,13 @@ class PyFMI_Wrapper:
         logger.debug('Successfully loaded model "%s"', model_name)
         return model
 
-    def setup(self, time_start, output_names, model_params):
+    def setup(self, time_start, output_names):
         self.model.reset()
         self.model.setup_experiment(start_time=time_start)
 
         # This is needed, because otherwise setting new values seems not to work
-        #self.model.enter_initialization_mode()
-        if model_params:
-            values = {var: f(time_start) for var, f in model_params.items()}
-            # list of keys and list of values
-            self.set_params(**values)
-        #self.model.exit_initialization_mode()
 
+        self.model.initialize()
         e_info = self.model.get_event_info()
         e_info.newDiscreteStatesNeeded = True
         # Event iteration
@@ -85,8 +80,6 @@ class PyFMI_Wrapper:
         self.model.set(*zip(*kwargs.items()))
 
     def set_params(self, **kwargs):
-        #self.model.enter_initialization_mode()
-        self.model.initialize()                 # replacing enter and exit mode -> works to set parameters during
-        # simulation AND model get outputs
+        self.model.initialize()
         self.model.set(*zip(*kwargs.items()))
-        #self.model.exit_initialization_mode()
+
