### 2020-08-28 a70b48f
Update openmodelica_microgrid_gym/agents/safeopt.py
