+++ openmodelica_microgrid_gym/env/pyfmi.py
@@ -27,8 +27,6 @@ class PyFMI_Wrapper:
         self.model.reset()
         self.model.setup_experiment(start_time=time_start)
 
-        # This is needed, because otherwise setting new values seems not to work
-
         self.model.initialize()
         e_info = self.model.get_event_info()
         e_info.newDiscreteStatesNeeded = True
@@ -82,4 +80,3 @@ class PyFMI_Wrapper:
     def set_params(self, **kwargs):
         self.model.initialize()
         self.model.set(*zip(*kwargs.items()))
-
