### 2020-08-28 #2435 @stheid
Dear @gonsie, all changes are now merged to master.
