### 2020-09-01 34eb5ed
#68 - added J to plot title
