+++ examples/single_inverter_current_control_safe_opt.py
@@ -7,7 +7,7 @@
 import logging
 import os
 from functools import partial
-from time import strftime, gmtime
+#from time import strftime, gmtime
 from typing import List
 
 import GPy
@@ -39,15 +39,16 @@ if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
 include_simulate = True
-show_plots = True
+show_plots = False
+balanced_load = False
 do_measurement = False
 
 # If True: Results are stored to directory mentioned in: REBASE to DEV after MERGE #60!!
-safe_results = False
+safe_results = True
 
 # Files saves results and  resulting plots to the folder saves_VI_control_safeopt in the current directory
 current_directory = os.getcwd()
-save_folder = os.path.join(current_directory, r'MCTest_bal')
+save_folder = os.path.join(current_directory, r'MC03_lenRun_unbal_lessnoise')
 os.makedirs(save_folder, exist_ok=True)
 
 lengthscale_vec = np.linspace(0.5, 0.5, 1)
@@ -58,8 +59,8 @@ np.random.seed(0)
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
 max_episode_steps = 1000  # number of simulation steps per episode
-num_episodes = 2  # number of simulation episodes (i.e. SafeOpt iterations)
-n_MC = 3  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
+num_episodes = 100  # number of simulation episodes (i.e. SafeOpt iterations)
+n_MC = 5  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
 # distribution to represent real world more accurate
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
@@ -75,8 +76,8 @@ i_ref = np.array([15, 0, 0])  # exemplary set point i.e. id = 15, iq = 0, i0 = 0
 
 # Controller layout due to magnitude optimum:
 L = 2.3e-3  # / H
-R = 400e-3#170e-3  # 585e-3  # / Ohm
-#R = 585e-3#170e-3  # 585e-3  # / Ohm
+R = 400e-3  # 170e-3  # 585e-3  # / Ohm
+# R = 585e-3#170e-3  # 585e-3  # / Ohm
 tau_plant = L / R
 gain_plant = 1 / R
 
@@ -85,15 +86,8 @@ Tn = tau_plant  # Due to compensate
 Kp_init = tau_plant / (2 * delta_t * gain_plant * v_DC)
 Ki_init = Kp_init / Tn
 
-#Kp_init = 0
 
-
-class pltManager:
-
-    def __init__(self, agent: Agent):
-        self.agent = agent
-
-
+# Kp_init = 0
 
 
 class Reward:
@@ -161,7 +155,7 @@ if __name__ == '__main__':
         # For 2D example, choose Kp and Ki as mutable parameters (below) and define bounds and lengthscale for both of them
         if adjust == 'Kpi':
             bounds = [(0.0, 4), (0, 200)]
-            lengthscale = [.3, 25.]
+            lengthscale = [0.3, 25.]
 
         # The performance should not drop below the safe threshold, which is defined by the factor safe_threshold times
         # the initial performance: safe_threshold = 0.8 means. Performance measurement for optimization are seen as
@@ -235,6 +229,36 @@ if __name__ == '__main__':
                              history=FullHistory()
                              )
 
+
+        class PlotManager:
+
[…]
