### 2020-09-01 ab9aee4
#68 - corrected performance cal of first run
