+++ openmodelica_microgrid_gym/execution/monte_carlo_runner.py
@@ -1,5 +1,4 @@
 import numpy as np
-import matplotlib.pyplot as plt
 from typing import Dict, Any
 from tqdm import tqdm
 from openmodelica_microgrid_gym.agents.episodic import EpisodicLearnerAgent
@@ -16,7 +15,7 @@ class MonteCarloRunner:
     Inside the MC-loop the observe function is called with terminated = False to only update the return.
     The return is stored in an array at the end of the MC-loop.
     After finishing the MC-loop, the average of the return-array is used to update the (controller) parameters.
-    Therefore, the agetn-observe function is called with terminated = True
+    Therefore, the agent-observe function is called with terminated = True
     """
 
     def __init__(self, agent: EpisodicLearnerAgent, env: ModelicaEnv):
@@ -37,19 +36,19 @@ class MonteCarloRunner:
         - "agent_plt": last agent plot
         """
 
-    def run(self, n_episodes: int = 10, n_MC: int = 5, visualise: bool = False, prepare_MC_experiment = lambda: True):
+    def run(self, n_episodes: int = 10, n_mc: int = 5, visualise: bool = False, prepare_mc_experiment=lambda: True):
         """
         Trains/executes the agent on the environment for a number of epochs
 
         :param n_episodes: number of epochs to play
-        :param n_MC: number of Monte-Carlo experiments using the same parameter set before updating the latter
+        :param n_mc: number of Monte-Carlo experiments using the same parameter set before updating the latter
         :param visualise: turns on visualization of the environment
         """
         self.agent.reset()
         self.env.history.cols = self.env.history.structured_cols(None) + self.agent.measurement_cols
         self.agent.obs_varnames = self.env.history.cols
 
-        performance_MC = np.zeros(n_MC)
+        performance_mc = np.zeros(n_mc)
 
         if not visualise:
             self.env.viz_mode = None
@@ -59,10 +58,9 @@ class MonteCarloRunner:
 
             done, r = False, None
             np.random.seed(0)
-            for m in tqdm(range(n_MC), desc='episodes', unit='epoch'):
+            for m in tqdm(range(n_mc), desc='episodes', unit='epoch'):
 
-
-                prepare_MC_experiment()
+                prepare_mc_experiment()
 
                 obs = self.env.reset()
 
@@ -73,10 +71,13 @@ class MonteCarloRunner:
                     obs, r, done, info = self.env.step(act)
                     self.env.render()
                     if done:
-                        self.agent.observe(r, False)  # take the last reward into account, too, but without update_params
-
+                        self.agent.observe(r,
+                                           False)  # take the last reward into account, too, but without update_params
+                        self.agent.performance = self.agent._iterations / (
+                                    self.agent.episode_return * self.agent.initial_performance)
                         if m == 0 and i == 0:
                             self.agent.initial_performance = self.agent.performance
+                            self.agent.performance = 1  # instead of perf/initial_perf
                             self.agent.last_best_performance = self.agent.performance
                             self.agent.last_worst_performance = self.agent.performance
 
@@ -85,11 +86,15 @@ class MonteCarloRunner:
                             self.agent.worst_episode = self.agent.history.df.shape[0]
                             self.agent.last_worst_performance = self.agent.performance
 
-                        performance_MC[m] = self.agent.performance
+                        performance_mc[m] = self.agent.performance
+                        # set iterations and episode return = 0
+                        self.agent.prepare_episode()
+
                         break
 
                 _, env_fig = self.env.close()
 
+                # vor break?
                 if (m == 0 and i == 0) or self.agent.has_improved:
                     self.run_data['best_env_plt'] = env_fig
                     self.run_data['best_episode_idx'] = i
@@ -100,24 +105,18 @@ class MonteCarloRunner:
[…]
