+++ examples/lengthScaleSweepMC.py
@@ -1,0 +1,489 @@
+#####################################
+# Example using a FMU by OpenModelica and SafeOpt algorithm to find optimal controller parameters
+# Simulation setup: Single inverter supplying 15 A d-current to an RL-load via a LC filter
+# Controller: PI current controller gain parameters are optimized by SafeOpt
+
+
+import logging
+import os
+from functools import partial
+#from time import strftime, gmtime
+from typing import List
+
+import GPy
+import gym
+import numpy as np
+
+import matplotlib.pyplot as plt
+
+from openmodelica_microgrid_gym import Agent
+from openmodelica_microgrid_gym.agents import SafeOptAgent
+from openmodelica_microgrid_gym.agents.util import MutableFloat
+from openmodelica_microgrid_gym.aux_ctl import PI_params, DroopParams, MultiPhaseDQCurrentSourcingController
+from openmodelica_microgrid_gym.env import PlotTmpl
+from openmodelica_microgrid_gym.env.physical_testbench import TestbenchEnv
+from openmodelica_microgrid_gym.env.stochastic_components import Load, Noise
+from openmodelica_microgrid_gym.execution.monte_carlo_runner import MonteCarloRunner
+from openmodelica_microgrid_gym.execution.runner_hardware import RunnerHardware
+from openmodelica_microgrid_gym.util import dq0_to_abc, nested_map, FullHistory
+
+# Choose which controller parameters should be adjusted by SafeOpt.
+# - Kp: 1D example: Only the proportional gain Kp of the PI controller is adjusted
+# - Ki: 1D example: Only the integral gain Ki of the PI controller is adjusted
+# - Kpi: 2D example: Kp and Ki are adjusted simultaneously
+
+adjust = 'Kpi'
+
+# Check if really only one simulation scenario was selected
+if adjust not in {'Kp', 'Ki', 'Kpi'}:
+    raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
+
+include_simulate = False
+show_plots = True
+balanced_load = False
+do_measurement = True
+
+# If True: Results are stored to directory mentioned in: REBASE to DEV after MERGE #60!!
+safe_results = False
+
+# Files saves results and  resulting plots to the folder saves_VI_control_safeopt in the current directory
+current_directory = os.getcwd()
+save_folder = os.path.join(current_directory, r'MC03_lenRun_unbal_lessnoise')
+os.makedirs(save_folder, exist_ok=True)
+
+lengthscale_vec = np.linspace(0.5, 0.5, 1)
+unsafe_vec = np.zeros(20)
+
+np.random.seed(0)
+
+# Simulation definitions
+delta_t = 1e-4  # simulation time step size / s
+max_episode_steps = 1000  # number of simulation steps per episode
+num_episodes = 100  # number of simulation episodes (i.e. SafeOpt iterations)
+n_MC = 5  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
+# distribution to represent real world more accurate
+v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
+nomFreq = 50  # nominal grid frequency / Hz
+nomVoltPeak = 230 * 1.414  # nominal grid voltage / V
+iLimit = 30  # inverter current limit / A
+iNominal = 20  # nominal inverter current / A
+mu = 2  # factor for barrier function (see below)
+DroopGain = 40000.0  # virtual droop gain for active power / W/Hz
[…]
