+++ examples/lengthScaleSweepMC.py
@@ -6,8 +6,9 @@
 
 import logging
 import os
+import pandas as pd
 from functools import partial
-#from time import strftime, gmtime
+# from time import strftime, gmtime
 from typing import List
 
 import GPy
@@ -38,29 +39,33 @@ adjust = 'Kpi'
 if adjust not in {'Kp', 'Ki', 'Kpi'}:
     raise ValueError("Please set 'adjust' to one of the following values: 'Kp', 'Ki', 'Kpi'")
 
-include_simulate = False
-show_plots = True
+include_simulate = True
+show_plots = False
 balanced_load = False
-do_measurement = True
+do_measurement = False
 
 # If True: Results are stored to directory mentioned in: REBASE to DEV after MERGE #60!!
 safe_results = False
 
 # Files saves results and  resulting plots to the folder saves_VI_control_safeopt in the current directory
 current_directory = os.getcwd()
-save_folder = os.path.join(current_directory, r'MC03_lenRun_unbal_lessnoise')
+save_folder = os.path.join(current_directory, r'lengthscaleSearchMC')
 os.makedirs(save_folder, exist_ok=True)
 
-lengthscale_vec = np.linspace(0.5, 0.5, 1)
-unsafe_vec = np.zeros(20)
+ii_steps = 1
+kk_steps = 10
+
+lengthscale_vec_kI = np.linspace(25, 25, ii_steps)
+lengthscale_vec_kP = np.linspace(0.1, 4, kk_steps)
+unsafe_vec = np.zeros([kk_steps, ii_steps])
 
 np.random.seed(0)
 
 # Simulation definitions
 delta_t = 1e-4  # simulation time step size / s
 max_episode_steps = 1000  # number of simulation steps per episode
-num_episodes = 100  # number of simulation episodes (i.e. SafeOpt iterations)
-n_MC = 5  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
+num_episodes = 50  # number of simulation episodes (i.e. SafeOpt iterations)
+n_MC = 20  # number of Monte-Carlo samples for simulation - samples device parameters (e.g. L,R, noise) from
 # distribution to represent real world more accurate
 v_DC = 60  # DC-link voltage / V; will be set as model parameter in the FMU
 nomFreq = 50  # nominal grid frequency / Hz
@@ -134,7 +139,7 @@ class Reward:
 
 if __name__ == '__main__':
 
-    for ll in range(len(lengthscale_vec)):
+    for kk in range(len(lengthscale_vec_kP)):
         #####################################
         # Definitions for the GP
         prior_mean = 0  # 2  # mean factor of the GP prior mean which is multiplied with the first performance of the initial set
@@ -155,7 +160,8 @@ if __name__ == '__main__':
         # For 2D example, choose Kp and Ki as mutable parameters (below) and define bounds and lengthscale for both of them
         if adjust == 'Kpi':
             bounds = [(0.0, 4), (0, 200)]
-            lengthscale = [0.3, 25.]
+            # lengthscale = [0.3, 25.]
+            lengthscale = [lengthscale_vec_kP[kk], 25.]
 
         # The performance should not drop below the safe threshold, which is defined by the factor safe_threshold times
         # the initial performance: safe_threshold = 0.8 means. Performance measurement for optimization are seen as
@@ -238,17 +244,17 @@ if __name__ == '__main__':
                 self.l_load = used_l_load
                 self.i_noise = used_i_noise
 
-                #self.r_load.gains =  [elem *1e3 for elem in self.r_load.gains]
-                #self.l_load.gains =  [elem *1e3 for elem in self.r_load.gains]
+                # self.r_load.gains =  [elem *1e3 for elem in self.r_load.gains]
[…]
