### 2020-09-02 bd2968d
#68 - monir lengthscale tests MC
