+++ examples/simple_agent.py
@@ -13,7 +13,7 @@ if __name__ == '__main__':
     env = gym.make('openmodelica_microgrid_gym:ModelicaEnv-v1',
                    model_input=['i1p1', 'i1p2', 'i1p3'],
                    model_output=dict(lc1=['inductor1.i', 'inductor2.i', 'inductor3.i']),
-                   model_path='../fmu/grid.network.fmu')
+                   model_path='../OpenModelica_Microgrids/grid.network.fmu')
 
     agent = RndAgent()
     runner = Runner(agent, env)
