### 2020-09-08 5737fbe
New library format + updating tests and paths in examples
