### 2020-09-13 #2435 @katyhuff
@dhhagan I have had a chance to review and I'm willing to accept, though I still have one concern:

Most of my comments were either optional or were handled by the authors. My concern about functionality was primarily regarding the clarity of examples and clarity of the functional claims in the paper. I made a comment previously that the functional claims of the software should be made more clear in the paper. The authors added a brief list of keywords to clarify what was meant by "auxiliaries," which helps. But, I still think the paper could benefit from even more clarity regarding exactly what is being claimed as functionality contributed by the work.
