### 2020-09-15 #2435 @stheid
@katyhuff Thanks for the additional clarification.

We had another discussion in the Team, but we find it difficult to add more than the bulleted list we have added. Therefore we would ask, if it would be possible to successfully close the review with the given state.
