### 2020-09-15 #2435 @dhhagan
@katyhuff Thanks for your feedback and review. @stheid: I would like to see a bit more information about the *many software auxiliaries* listed in the **Features** section. It doesn't currently tell me what is possible/available, which I think would be beneficial and help make the paper itself more clear. In addition, there are a few areas where the language could be cleaned up - I will send a PR with suggestions in a little bit and you can take/leave the ones you want and/or feel are necessary.
