### 2020-09-15 70822c8
Update paper.md
Made suggestions regarding text clarifications to make the reading a bit easier. Please feel free to accept/not accept any/all of them.
