+++ docs/JOSS/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-The OpenModelica Microgrid Gym (OMG) toolbox is providing a transient simulation framework for local energy grids based on power electronic converters. OpenModelica is used in the backend allowing to set up arbitrary electric grid designs using its well-known graphical user interface in a plug & play fashion [@Fritzson2018]. A Python-interface is offered for configuring the simulations and specially to integrate software modules for the realization as well testing of closed control loops. In addition, the OpenAI Gym interface is provided for connecting data-driven reinforcement learning algorithms in order to investigate on intelligent microgrid control approaches [@OpenAI2016]. 
+The OpenModelica Microgrid Gym (OMG) toolbox provides a transient simulation framework for local energy grids based on power electronic converters. OpenModelica is used as the backend, allowing users to set up arbitrary electric grid designs via its well-known graphical user interface in a plug-and-play fashion [@Fritzson2018]. Simulations can be configured using a python interface, making it easy to integrate software modules for the realization and testing of closed control loops. In addition, the OpenAI Gym interface is provided to connect data-driven reinforcement learning algorithms for investigating intelligent microgrid control approaches [@OpenAI2016]. 
 
 ![\label{fig:omg}](omg.png)
 _Fig. 1:  Overview of the interconnections between the different parts of the  OMG  toolbox.  The  OpenModelica and  OpenAIGym logos are the property of their respective owners._
@@ -40,35 +40,35 @@ _Fig. 1:  Overview of the interconnections between the different parts of the  O
 
 # Background on microgrids and their control
 
-Micro- and smart grids (MSG) play an important role both for integrating renewable energy sources in conventional electricity grids and for providing power supply in remote areas [@Lund2017]. 
+Micro- and smart-grids (MSG) play an important role for integrating renewable energy sources in conventional electricity grids and for providing power supply in remote areas [@Lund2017]. 
 Due to their high efficiency and flexibility, power electronic converters are largely used to drive modern MSG.
-Power electronics is the application of solid-state electronics to the control and conversion of electric power which is largely performed with semiconductor switching devices such as diodes or power transistors.
-This includes energy conversion in terms of voltage and current amplitude, frequency and phase angle as well as the number of phases between two or more electrical energy systems to be connected.
+Power electronics describes the application of solid-state electronics to the control and conversion of electric power, which is largely performed with semiconductor switching devices such as diodes or power transistors.
+This includes energy conversion in terms of voltage and current amplitude, frequency and phase angle, as well as the number of phases between two or more electrical energy systems to be connected.
 
 
-Controlling MSGs is a challenging task due to the high requirements on energy availability, safety and voltage quality. 
+Controlling MSGs is a challenging task due to the high requirements on energy availability, safety, and voltage quality. 
 This is particularly demanding due to the wide range of different MSG topologies depending on their field of application like industrial campuses, residential areas or remote off-grid electrification [@Kroposki2008].
 This results in high demand for comprehensive testing of new control concepts during their development phase and comparisons with the state of the art to ensure their feasibility.
 This applies in particular to data-driven control approaches such as reinforcement learning (RL), the stability and operating behavior of which cannot be evaluated a priori [@Garcia2015].
 
 
 # State of field 
-``OMG`` is a Python-based package for modeling and simulation of microgrids based on power electronic energy conversion.
-An attached OpenModelica [@Fritzson2018] library enables the user to define their microgrid (i.e. a local electricity grid containing arbitrary sources, storages and loads) in a flexible and scalable way or to use predefined example grids. 
[…]
