### 2020-09-15 #2435 @dhhagan
@stheid I've made a PR (https://github.com/upb-lea/openmodelica-microgrid-gym/pull/81) with some suggested text changes in addition to what was mentioned above about expanding the clarity of examples and *many software auxiliaries* listed above. Please let me know when you've had time to review these and I'll take another look.
