### 2020-09-16 4d9ec14
New library format + updating tests and paths in examples (#80)
* New library format + updating tests and paths in examples

* Updating the Documentation, especially the parts connected with the Library

* Adding one more example + components to the Library
Actualizing the pictures


Co-authored-by: Daniel Weber <weber@lea.uni-paderborn.de>
