+++ examples/basic_env_norm.py
@@ -3,7 +3,7 @@ import gym
 if __name__ == '__main__':
     env = gym.make('openmodelica_microgrid_gym:NormalizedEnv_test-v1',
                    net='net.yaml',
-                   model_path='../fmu/grid.network.fmu')
+                   model_path='../OpenModelica_Microgrids/OpenModelica_Microgrids.Grids.Network.fmu')
 
     env.reset()
     for _ in range(1000):
