### 2020-09-17 #2435 @stheid
Thank you very much for the PR, it is already merged :)

Additionally, we clarified this paragraph further. https://github.com/upb-lea/openmodelica-microgrid-gym/commit/de90dc9a74c1bcbe2581f7d680cfa2879ea9f6e6
