### 2020-09-18 c812be7
#84 - fixes bug that safeopt plots are only shown for the first couple of iterations (#86)
Works in examples
