### 2020-09-21 #2435 @stheid
It seems like Zenodo forces us to make a new release in order to get a DOI. We are currently discussing it we make a small release just for this technical reasons or if we wait for the master to be in release worthy shape.
