### 2020-09-21 #2435 @stheid
This should be the DOI https://doi.org/10.5281/zenodo.4041278

Zenodo seems to have issues in their infrastructure right now, because the creation of this DOI took more than 4 hours. Therefore i think it should not consern that the DOI cannot be resolved yet. Hopefully it will work tomorrow.
