### 2020-09-22 #2435 @dhhagan
@stheid It looks like the Zenodo archive doesn't contain the full/correct author list. Could you correct that and then post the final OMG version and DOI archive? As soon as that's updated, I can accept this manuscript. 👍
