### 2020-09-23 #2435 @stheid
Oh, thanks a lot. I completely overlooked that i have to edit the entry to be consistent. Its corrected now.

I think the DOI did not change. As said currently the versions state is for the technical reason in a bit of an unfortunate state, and we are not ready to roll out a new release right now. So we would keep everything at 0.2.0-2 if this is ok.

Where is the DOI of the release connected to JOSS. Do we need to update the something for future releases, because i assume after the reviewing process here is finished, the project will still continue to grow.
