### 2020-09-24 #2435 @arfon
> I think the DOI did not change. As said currently the versions state is for the technical reason in a bit of an unfortunate state, and we are not ready to roll out a new release right now. So we would keep everything at 0.2.0-2 if this is ok.

The thing we (JOSS) care about here is that the archive of the software includes the changes that have resulted from this review.

> Where is the DOI of the release connected to JOSS. Do we need to update the something for future releases, because i assume after the reviewing process here is finished, the project will still continue to grow.

We reference this archive in the JOSS metadata and it's linked to from the paper and the website.
