### 2020-09-24 #2435 @stheid
> The thing we (JOSS) care about here is that the archive of the software includes the changes that have resulted from this review.

that is the case! :)

> We reference this archive in the JOSS metadata and it's linked to from the paper and the website.

In that case maybe using the version-agnostic DOI is better? 10.5281/zenodo.4041277
But in any case if you are fine with everything, we are too.

Thanks a lot for clearing things up.
