### 2020-09-24 #2435 @arfon
I think we would prefer to link to `10.5281/zenodo.4041278` as this is the version most closely associated with the review. Don't forget that the paper also links to the software repository in the margin too so readers will have a link to the live software repository to access the latest version of the code.
