### 2020-10-01 #2435 @Kevin-Mattheus-Moerman
@dhhagan thanks for editing this one. I'll process this now for acceptance in JOSS. One point, for future reference, please always check that the ZENODO meta data (authors, title, version tag) match that of the paper.
