### 2020-10-01 #2435 @Kevin-Mattheus-Moerman
@stheid I'll be processing this work for acceptance in JOSS. Please work on the remaining minor points below, thanks:

- [ ] The reviewed submission is at `v0.1.3` while the provided archived version is at `v0.2.0-2`. Please create a new archived version for `v0.1.3`. When you do this also make sure the ZENODO meta-data for the author names and the title match as well. **You may need to manually edit this.**

- [ ] Please add city and country information for all affiliations (do not use acronyms for countries)

- [ ] This work is about to be accepted in JOSS. We recommend you proofread your paper once more yourself at this stage, and also verify the accuracy of the author names, affiliations, and acknowledgments.
