### 2020-10-01 #2435 @stheid
That all sounds good :)

@Kevin-Mattheus-Moerman How to create a zenodo archive of a already released version. I have the impression this is impossible (that was the only reason we created 0.2.0-2 in the first place)
