### 2020-10-02 #2435 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
