### 2020-10-03 #2435 @Kevin-Mattheus-Moerman
@stheid please edit the affiliation in the YAML at the top of your `paper.md` file.
