### 2020-10-05 #2435 @Kevin-Mattheus-Moerman
> Like so [upb-lea/openmodelica-microgrid-gym@6eeabca](https://github.com/upb-lea/openmodelica-microgrid-gym/commit/6eeabca6163c399f8ab494f57c297d3235b81fbf)?

@stheid yes but please also add the city
