### 2020-10-05 a259b40
Update paper.md
added city to paper authors affiliations
