### 2020-10-05 #2435 @stheid
https://github.com/upb-lea/openmodelica-microgrid-gym/commit/a259b405cf4760bc3af1979f6d51b4ce2e00c555

Sorry, i thought it would be clear enough ^^;  But i added it now.
